MINOR: Split up ShareConsumerTest
