+++ clients/clients-integration-tests/src/test/java/org/apache/kafka/clients/consumer/ShareConsumerTest.java
@@ -19,20 +19,7 @@ package org.apache.kafka.clients.consumer;
 import kafka.server.KafkaBroker;
 
 import org.apache.kafka.clients.admin.Admin;
-import org.apache.kafka.clients.admin.AlterConfigOp;
-import org.apache.kafka.clients.admin.AlterConfigsOptions;
-import org.apache.kafka.clients.admin.AlterShareGroupOffsetsOptions;
-import org.apache.kafka.clients.admin.Config;
-import org.apache.kafka.clients.admin.ConfigEntry;
-import org.apache.kafka.clients.admin.CreateTopicsResult;
-import org.apache.kafka.clients.admin.DeleteShareGroupOffsetsOptions;
-import org.apache.kafka.clients.admin.DescribeShareGroupsOptions;
-import org.apache.kafka.clients.admin.ListShareGroupOffsetsOptions;
-import org.apache.kafka.clients.admin.ListShareGroupOffsetsResult;
-import org.apache.kafka.clients.admin.ListShareGroupOffsetsSpec;
-import org.apache.kafka.clients.admin.NewTopic;
 import org.apache.kafka.clients.admin.RecordsToDelete;
-import org.apache.kafka.clients.admin.ShareMemberDescription;
 import org.apache.kafka.clients.admin.SharePartitionOffsetInfo;
 import org.apache.kafka.clients.producer.Producer;
 import org.apache.kafka.clients.producer.ProducerConfig;
@@ -42,11 +29,8 @@ import org.apache.kafka.common.KafkaException;
 import org.apache.kafka.common.TopicIdPartition;
 import org.apache.kafka.common.TopicPartition;
 import org.apache.kafka.common.Uuid;
-import org.apache.kafka.common.config.ConfigResource;
 import org.apache.kafka.common.errors.GroupMaxSizeReachedException;
 import org.apache.kafka.common.errors.InterruptException;
-import org.apache.kafka.common.errors.InvalidConfigurationException;
-import org.apache.kafka.common.errors.InvalidRecordStateException;
 import org.apache.kafka.common.errors.InvalidTopicException;
 import org.apache.kafka.common.errors.NotLeaderOrFollowerException;
 import org.apache.kafka.common.errors.RecordDeserializationException;
@@ -57,7 +41,6 @@ import org.apache.kafka.common.errors.WakeupException;
 import org.apache.kafka.common.header.Header;
 import org.apache.kafka.common.header.Headers;
 import org.apache.kafka.common.internals.Topic;
-import org.apache.kafka.common.network.ListenerName;
 import org.apache.kafka.common.record.TimestampType;
 import org.apache.kafka.common.serialization.ByteArrayDeserializer;
 import org.apache.kafka.common.serialization.Deserializer;
@@ -71,18 +54,10 @@ import org.apache.kafka.common.test.api.ClusterTestDefaults;
 import org.apache.kafka.common.test.api.ClusterTests;
 import org.apache.kafka.common.test.api.Type;
 import org.apache.kafka.common.utils.Utils;
-import org.apache.kafka.coordinator.group.GroupConfig;
 import org.apache.kafka.coordinator.group.GroupCoordinatorConfig;
-import org.apache.kafka.coordinator.group.modern.share.ShareGroupConfig;
-import org.apache.kafka.server.metrics.KafkaYammerMetrics;
 import org.apache.kafka.server.share.SharePartitionKey;
-import org.apache.kafka.test.NoRetryException;
 import org.apache.kafka.test.TestUtils;
 
-import com.yammer.metrics.core.Meter;
-
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Timeout;
 
@@ -90,7 +65,6 @@ import java.nio.ByteBuffer;
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collection;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Iterator;
@@ -98,23 +72,17 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
-import java.util.Properties;
 import java.util.Set;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
-import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
-import java.util.concurrent.Future;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicReference;
-import java.util.function.BiConsumer;
 import java.util.function.Function;
-import java.util.function.Predicate;
-import java.util.stream.IntStream;
 import java.util.stream.Stream;
 
 import static org.apache.kafka.test.TestUtils.DEFAULT_MAX_WAIT_MS;
@@ -131,7 +99,6 @@ import static org.junit.jupiter.api.Assertions.assertTimeoutPreemptively;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.fail;
 
-@SuppressWarnings({"ClassFanOutComplexity", "ClassDataAbstractionCoupling"})
 @Timeout(1200)
 @Tag("integration")
 @ClusterTestDefaults(
@@ -149,42 +116,10 @@ import static org.junit.jupiter.api.Assertions.fail;
         @ClusterConfigProperty(key = "transaction.state.log.replication.factor", value = "1")
     }
 )
-public class ShareConsumerTest {
-    private final ClusterInstance cluster;
-    private final TopicPartition tp = new TopicPartition("topic", 0);
-    private Uuid tpId;
-    private final TopicPartition tp2 = new TopicPartition("topic2", 0);
-    private final TopicPartition warmupTp = new TopicPartition("warmup", 0);
-    private List<TopicPartition> sgsTopicPartitions;
-    private static final String KEY = "content-type";
-    private static final String VALUE = "application/octet-stream";
-    private static final String EXPLICIT = "explicit";
-    private static final String IMPLICIT = "implicit";
-    private static final String RECORD_LIMIT = "record_limit";
-    private static final String BATCH_OPTIMIZED = "batch_optimized";
+public class ShareConsumerTest extends ShareConsumerTestBase {
 
     public ShareConsumerTest(ClusterInstance cluster) {
-        this.cluster = cluster;
-    }
-
-    @BeforeEach
-    public void setup() {
-        try {
-            this.cluster.waitForReadyBrokers();
-            tpId = createTopic("topic");
-            createTopic("topic2");
-            sgsTopicPartitions = IntStream.range(0, 3)
-                .mapToObj(part -> new TopicPartition(Topic.SHARE_GROUP_STATE_TOPIC_NAME, part))
-                .toList();
-            this.warmup();
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @AfterEach
-    public void tearDown() {
-        kafka.utils.TestUtils.clearYammerMetrics();
+        super(cluster);
     }
 
     @ClusterTest
@@ -277,10 +212,10 @@ public class ShareConsumerTest {
 
     @ClusterTests({
         @ClusterTest(serverProperties = {
-            @ClusterConfigProperty(key = GroupCoordinatorConfig.CONSUMER_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "0")
+            @ClusterConfigProperty(key = GroupCoordinatorConfig.SHARE_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "0")
         }),
         @ClusterTest(serverProperties = {
-            @ClusterConfigProperty(key = GroupCoordinatorConfig.CONSUMER_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "1000")
+            @ClusterConfigProperty(key = GroupCoordinatorConfig.SHARE_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "1000")
         })
     })
     public void testSubscriptionAndPoll() {
@@ -302,10 +237,10 @@ public class ShareConsumerTest {
 
     @ClusterTests({
         @ClusterTest(serverProperties = {
-            @ClusterConfigProperty(key = GroupCoordinatorConfig.CONSUMER_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "0")
+            @ClusterConfigProperty(key = GroupCoordinatorConfig.SHARE_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "0")
         }),
         @ClusterTest(serverProperties = {
-            @ClusterConfigProperty(key = GroupCoordinatorConfig.CONSUMER_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "1000")
+            @ClusterConfigProperty(key = GroupCoordinatorConfig.SHARE_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "1000")
         })
     })
     public void testSubscriptionAndPollMultiple() {
@@ -346,196 +281,6 @@ public class ShareConsumerTest {
             ConsumerRecords<byte[], byte[]> records = waitedPoll(shareConsumer, 2500L, 1);
             assertEquals(1, records.count());
             verifyShareGroupStateTopicRecordsProduced();
-        }
-    }
-
-    @ClusterTest
-    public void testAcknowledgementSentOnSubscriptionChange() throws ExecutionException, InterruptedException {
-        alterShareAutoOffsetReset("group1", "earliest");
-        try (Producer<byte[], byte[]> producer = createProducer();
-             ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer("group1")) {
-
-            Map<TopicPartition, Set<Long>> partitionOffsetsMap = new HashMap<>();
-            Map<TopicPartition, Exception> partitionExceptionMap = new HashMap<>();
-
-            ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(), null, "key".getBytes(), "value".getBytes());
-            producer.send(record);
-            ProducerRecord<byte[], byte[]> record2 = new ProducerRecord<>(tp2.topic(), tp2.partition(), null, "key".getBytes(), "value".getBytes());
-            producer.send(record2).get();
-            producer.flush();
-            shareConsumer.setAcknowledgementCommitCallback(new TestableAcknowledgementCommitCallback(partitionOffsetsMap, partitionExceptionMap));
-
-            shareConsumer.subscribe(Set.of(tp.topic()));
-            assertEquals(Optional.empty(), shareConsumer.acquisitionLockTimeoutMs());
-
-            TestUtils.waitForCondition(() -> shareConsumer.poll(Duration.ofMillis(2000)).count() == 1,
-                DEFAULT_MAX_WAIT_MS, 100L, () -> "Failed to consume records for share consumer");
-
-            assertEquals(Optional.of(15000), shareConsumer.acquisitionLockTimeoutMs());
-            shareConsumer.subscribe(Set.of(tp2.topic()));
-            assertEquals(Optional.of(15000), shareConsumer.acquisitionLockTimeoutMs());
-
-            // Waiting for heartbeat to propagate the subscription change.
-            TestUtils.waitForCondition(() -> {
-                shareConsumer.poll(Duration.ofMillis(500));
-                return partitionOffsetsMap.containsKey(tp) && partitionOffsetsMap.containsKey(tp2);
-            }, DEFAULT_MAX_WAIT_MS, 100L, () -> "Failed to consume records from the updated subscription");
-
-            // Verifying if the callback was invoked without exceptions for the partitions for both topics.
-            assertFalse(partitionExceptionMap.containsKey(tp));
-            assertFalse(partitionExceptionMap.containsKey(tp2));
-            verifyShareGroupStateTopicRecordsProduced();
-        }
-    }
-
-    @ClusterTest
-    public void testAcknowledgementCommitCallbackSuccessfulAcknowledgement() throws Exception {
-        alterShareAutoOffsetReset("group1", "earliest");
-        try (Producer<byte[], byte[]> producer = createProducer();
-             ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer("group1")) {
-
-            Map<TopicPartition, Set<Long>> partitionOffsetsMap = new HashMap<>();
-            Map<TopicPartition, Exception> partitionExceptionMap = new HashMap<>();
-            ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(), null, "key".getBytes(), "value".getBytes());
-
-            producer.send(record);
-            producer.flush();
-
-            shareConsumer.setAcknowledgementCommitCallback(new TestableAcknowledgementCommitCallback(partitionOffsetsMap, partitionExceptionMap));
-            shareConsumer.subscribe(Set.of(tp.topic()));
-
-            TestUtils.waitForCondition(() -> shareConsumer.poll(Duration.ofMillis(2000)).count() == 1,
-                DEFAULT_MAX_WAIT_MS, 100L, () -> "Failed to consume records for share consumer");
-
-            // The callback should be called before the return of the poll, even when there are no more records.
-            ConsumerRecords<byte[], byte[]> records = shareConsumer.poll(Duration.ofMillis(2000));
-            assertEquals(0, records.count());
-            assertTrue(partitionOffsetsMap.containsKey(tp));
-
-            // We expect no exception as the acknowledgement error code is null.
-            assertFalse(partitionExceptionMap.containsKey(tp));
-            verifyShareGroupStateTopicRecordsProduced();
-        }
-    }
-
-    @ClusterTest
-    public void testAcknowledgementCommitCallbackSuccessfulAcknowledgementOnCommitSync() throws Exception {
-        alterShareAutoOffsetReset("group1", "earliest");
-        try (Producer<byte[], byte[]> producer = createProducer();
-             ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer("group1")) {
-
-            Map<TopicPartition, Set<Long>> partitionOffsetsMap = new HashMap<>();
-            Map<TopicPartition, Exception> partitionExceptionMap = new HashMap<>();
-            ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(), null, "key".getBytes(), "value".getBytes());
-
-            producer.send(record);
-            producer.flush();
-
-            shareConsumer.setAcknowledgementCommitCallback(new TestableAcknowledgementCommitCallback(partitionOffsetsMap, partitionExceptionMap));
-            shareConsumer.subscribe(Set.of(tp.topic()));
-
-            TestUtils.waitForCondition(() -> shareConsumer.poll(Duration.ofMillis(2000)).count() == 1,
-                DEFAULT_MAX_WAIT_MS, 100L, () -> "Failed to consume records for share consumer");
-
-            // The acknowledgement commit callback should be called before the commitSync returns
-            // once the records have been confirmed to have been acknowledged.
-            Map<TopicIdPartition, Optional<KafkaException>> result = shareConsumer.commitSync();
-            assertEquals(1, result.size());
-            assertTrue(partitionOffsetsMap.containsKey(tp));
-
-            // We expect no exception as the acknowledgement error code is null.
-            assertFalse(partitionExceptionMap.containsKey(tp));
-            verifyShareGroupStateTopicRecordsProduced();
-        }
-    }
-
-    @ClusterTest
-    public void testAcknowledgementCommitCallbackOnClose() {
-        alterShareAutoOffsetReset("group1", "earliest");
-        try (Producer<byte[], byte[]> producer = createProducer();
-             ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer("group1")) {
-
-            Map<TopicPartition, Set<Long>> partitionOffsetsMap = new HashMap<>();
-            Map<TopicPartition, Exception> partitionExceptionMap = new HashMap<>();
-            ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(), null, "key".getBytes(), "value".getBytes());
-            producer.send(record);
-            producer.flush();
-            shareConsumer.setAcknowledgementCommitCallback(new TestableAcknowledgementCommitCallback(partitionOffsetsMap, partitionExceptionMap));
-            shareConsumer.subscribe(Set.of(tp.topic()));
-
-            ConsumerRecords<byte[], byte[]> records = waitedPoll(shareConsumer, 2500L, 1);
-            assertEquals(1, records.count());
-
-            // Now in the second poll, we implicitly acknowledge the record received in the first poll.
-            // We get back the acknowledgement error code asynchronously after the second poll.
-            // The acknowledgement commit callback is invoked in close.
-            shareConsumer.poll(Duration.ofMillis(1000));
-            shareConsumer.close();
-
-            // We expect no exception as the acknowledgement error code is null.
-            assertFalse(partitionExceptionMap.containsKey(tp));
-            verifyShareGroupStateTopicRecordsProduced();
-        }
-    }
-
-    @ClusterTest
-    public void testAcknowledgementCommitCallbackInvalidRecordStateException() throws Exception {
-        alterShareAutoOffsetReset("group1", "earliest");
-        try (Producer<byte[], byte[]> producer = createProducer();
-             ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer("group1")) {
-
-            Map<TopicPartition, Set<Long>> partitionOffsetsMap = new HashMap<>();
-            Map<TopicPartition, Exception> partitionExceptionMap = new HashMap<>();
-            ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(), null, "key".getBytes(), "value".getBytes());
-            producer.send(record);
-            producer.flush();
-
-            shareConsumer.setAcknowledgementCommitCallback(new TestableAcknowledgementCommitCallback(partitionOffsetsMap, partitionExceptionMap));
-            shareConsumer.subscribe(Set.of(tp.topic()));
-
-            ConsumerRecords<byte[], byte[]> records = waitedPoll(shareConsumer, 2500L, 1);
-            assertEquals(1, records.count());
-
-            // Waiting until the acquisition lock expires.
-            Thread.sleep(20000);
-
-            TestUtils.waitForCondition(() -> {
-                shareConsumer.poll(Duration.ofMillis(500));
-                return partitionExceptionMap.containsKey(tp) && partitionExceptionMap.get(tp) instanceof InvalidRecordStateException;
-            }, DEFAULT_MAX_WAIT_MS, 100L, () -> "Failed to be notified by InvalidRecordStateException");
-        }
-    }
-
-    /**
-     * Test implementation of AcknowledgementCommitCallback to track the completed acknowledgements.
-     * partitionOffsetsMap is used to track the offsets acknowledged for each partition.
-     * partitionExceptionMap is used to track the exception encountered for each partition if any.
-     * Note - Multiple calls to {@link #onComplete(Map, Exception)} will not update the partitionExceptionMap for any existing partitions,
-     * so please ensure to clear the partitionExceptionMap after every call to onComplete() in a single test.
-     */
-    private static class TestableAcknowledgementCommitCallback implements AcknowledgementCommitCallback {
-        private final Map<TopicPartition, Set<Long>> partitionOffsetsMap;
-        private final Map<TopicPartition, Exception> partitionExceptionMap;
-
-        public TestableAcknowledgementCommitCallback(Map<TopicPartition, Set<Long>> partitionOffsetsMap,
-                                                     Map<TopicPartition, Exception> partitionExceptionMap) {
-            this.partitionOffsetsMap = partitionOffsetsMap;
-            this.partitionExceptionMap = partitionExceptionMap;
-        }
-
-        @Override
-        public void onComplete(Map<TopicIdPartition, Set<Long>> offsetsMap, Exception exception) {
-            offsetsMap.forEach((partition, offsets) -> {
-                partitionOffsetsMap.merge(partition.topicPartition(), offsets, (oldOffsets, newOffsets) -> {
-                    Set<Long> mergedOffsets = new HashSet<>();
-                    mergedOffsets.addAll(oldOffsets);
-                    mergedOffsets.addAll(newOffsets);
-                    return mergedOffsets;
-                });
-                if (!partitionExceptionMap.containsKey(partition.topicPartition()) && exception != null) {
-                    partitionExceptionMap.put(partition.topicPartition(), exception);
-                }
-            });
         }
     }
 
@@ -1444,10 +1189,10 @@ public class ShareConsumerTest {
 
     @ClusterTests({
         @ClusterTest(serverProperties = {
-            @ClusterConfigProperty(key = GroupCoordinatorConfig.CONSUMER_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "0")
+            @ClusterConfigProperty(key = GroupCoordinatorConfig.SHARE_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "0")
         }),
         @ClusterTest(serverProperties = {
-            @ClusterConfigProperty(key = GroupCoordinatorConfig.CONSUMER_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "1000")
+            @ClusterConfigProperty(key = GroupCoordinatorConfig.SHARE_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "1000")
         })
     })
     public void testMultipleConsumersInGroupSequentialConsumption() {
@@ -1487,10 +1232,10 @@ public class ShareConsumerTest {
 
     @ClusterTests({
         @ClusterTest(serverProperties = {
-            @ClusterConfigProperty(key = GroupCoordinatorConfig.CONSUMER_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "0")
+            @ClusterConfigProperty(key = GroupCoordinatorConfig.SHARE_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "0")
         }),
         @ClusterTest(serverProperties = {
-            @ClusterConfigProperty(key = GroupCoordinatorConfig.CONSUMER_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "1000")
+            @ClusterConfigProperty(key = GroupCoordinatorConfig.SHARE_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "1000")
         })
     })
     public void testMultipleConsumersInGroupConcurrentConsumption()
@@ -1528,10 +1273,10 @@ public class ShareConsumerTest {
 
     @ClusterTests({
         @ClusterTest(serverProperties = {
-            @ClusterConfigProperty(key = GroupCoordinatorConfig.CONSUMER_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "0")
+            @ClusterConfigProperty(key = GroupCoordinatorConfig.SHARE_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "0")
         }),
         @ClusterTest(serverProperties = {
-            @ClusterConfigProperty(key = GroupCoordinatorConfig.CONSUMER_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "1000")
+            @ClusterConfigProperty(key = GroupCoordinatorConfig.SHARE_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "1000")
         })
     })
     public void testMultipleConsumersInMultipleGroupsConcurrentConsumption()
@@ -1648,10 +1393,10 @@ public class ShareConsumerTest {
 
     @ClusterTests({
         @ClusterTest(serverProperties = {
-            @ClusterConfigProperty(key = GroupCoordinatorConfig.CONSUMER_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "0")
+            @ClusterConfigProperty(key = GroupCoordinatorConfig.SHARE_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "0")
         }),
         @ClusterTest(serverProperties = {
-            @ClusterConfigProperty(key = GroupCoordinatorConfig.CONSUMER_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "1000")
+            @ClusterConfigProperty(key = GroupCoordinatorConfig.SHARE_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "1000")
         })
     })
     public void testMultipleConsumersInGroupFailureConcurrentConsumption()
@@ -1759,148 +1504,6 @@ public class ShareConsumerTest {
     }
 
     /**
-     * Test to verify that the acknowledgement commit callback cannot invoke methods of ShareConsumer.
-     * The exception thrown is verified in {@link TestableAcknowledgementCommitCallbackWithShareConsumer}
-     */
-    @ClusterTest
-    public void testAcknowledgementCommitCallbackCallsShareConsumerDisallowed() {
-        alterShareAutoOffsetReset("group1", "earliest");
-        try (Producer<byte[], byte[]> producer = createProducer();
-             ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer("group1")) {
-
-            ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(), null, "key".getBytes(), "value".getBytes());
-            producer.send(record);
-            producer.flush();
-
-            shareConsumer.setAcknowledgementCommitCallback(new TestableAcknowledgementCommitCallbackWithShareConsumer<>(shareConsumer));
-            shareConsumer.subscribe(Set.of(tp.topic()));
-
-            // The acknowledgement commit callback will try to call a method of ShareConsumer
-            shareConsumer.poll(Duration.ofMillis(5000));
-            // The second poll sends the acknowledgements implicitly.
-            // The acknowledgement commit callback will be called and the exception is thrown.
-            // This is verified inside the onComplete() method implementation.
-            shareConsumer.poll(Duration.ofMillis(500));
-            verifyShareGroupStateTopicRecordsProduced();
-        }
-    }
-
-    private class TestableAcknowledgementCommitCallbackWithShareConsumer<K, V> implements AcknowledgementCommitCallback {
-        private final ShareConsumer<K, V> shareConsumer;
-
-        TestableAcknowledgementCommitCallbackWithShareConsumer(ShareConsumer<K, V> shareConsumer) {
-            this.shareConsumer = shareConsumer;
-        }
-
-        @Override
-        public void onComplete(Map<TopicIdPartition, Set<Long>> offsetsMap, Exception exception) {
-            // Accessing methods of ShareConsumer should throw an exception.
-            assertThrows(IllegalStateException.class, shareConsumer::close);
-            assertThrows(IllegalStateException.class, () -> shareConsumer.subscribe(Set.of(tp.topic())));
-            assertThrows(IllegalStateException.class, () -> shareConsumer.poll(Duration.ofMillis(5000)));
-        }
-    }
-
-    /**
-     * Test to verify that the acknowledgement commit callback can invoke ShareConsumer.wakeup() and it
-     * wakes up the enclosing poll.
-     */
-    @ClusterTest
-    public void testAcknowledgementCommitCallbackCallsShareConsumerWakeup() throws InterruptedException {
-        alterShareAutoOffsetReset("group1", "earliest");
-        try (Producer<byte[], byte[]> producer = createProducer();
-             ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer("group1")) {
-
-            ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(), null, "key".getBytes(), "value".getBytes());
-            producer.send(record);
-            producer.flush();
-
-            // The acknowledgement commit callback will try to call a method of ShareConsumer
-            shareConsumer.setAcknowledgementCommitCallback(new TestableAcknowledgementCommitCallbackWakeup<>(shareConsumer));
-            shareConsumer.subscribe(Set.of(tp.topic()));
-
-            TestUtils.waitForCondition(() -> shareConsumer.poll(Duration.ofMillis(2000)).count() == 1,
-                DEFAULT_MAX_WAIT_MS, 100L, () -> "Failed to consume records for share consumer");
-
-            // The second poll sends the acknowledgements implicitly.
-            shareConsumer.poll(Duration.ofMillis(2000));
-
-            // Till now acknowledgement commit callback has not been called, so no exception thrown yet.
-            // On 3rd poll, the acknowledgement commit callback will be called and the exception is thrown.
-            AtomicBoolean exceptionThrown = new AtomicBoolean(false);
-            TestUtils.waitForCondition(() -> {
-                try {
-                    shareConsumer.poll(Duration.ofMillis(500));
-                } catch (WakeupException e) {
-                    exceptionThrown.set(true);
-                }
-                return exceptionThrown.get();
-            }, DEFAULT_MAX_WAIT_MS, 100L, () -> "Failed to receive expected exception");
-            verifyShareGroupStateTopicRecordsProduced();
-        }
-    }
-
-    private static class TestableAcknowledgementCommitCallbackWakeup<K, V> implements AcknowledgementCommitCallback {
-        private final ShareConsumer<K, V> shareConsumer;
-
-        TestableAcknowledgementCommitCallbackWakeup(ShareConsumer<K, V> shareConsumer) {
-            this.shareConsumer = shareConsumer;
-        }
-
-        @Override
-        public void onComplete(Map<TopicIdPartition, Set<Long>> offsetsMap, Exception exception) {
-            shareConsumer.wakeup();
-        }
-    }
-
-    /**
-     * Test to verify that the acknowledgement commit callback can throw an exception, and it is not propagated
-     * to the caller of poll().
-     */
-    @ClusterTest
-    public void testAcknowledgementCommitCallbackThrowsException() throws InterruptedException {
-        alterShareAutoOffsetReset("group1", "earliest");
-        try (Producer<byte[], byte[]> producer = createProducer();
-             ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer("group1")) {
-
-            ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(), null, "key".getBytes(), "value".getBytes());
-            producer.send(record);
-            producer.flush();
-
-            AtomicBoolean callbackCalled = new AtomicBoolean(false);
-            shareConsumer.setAcknowledgementCommitCallback(new TestableAcknowledgementCommitCallbackThrows(callbackCalled));
-            shareConsumer.subscribe(Set.of(tp.topic()));
-
-            TestUtils.waitForCondition(() -> shareConsumer.poll(Duration.ofMillis(2000)).count() == 1,
-                DEFAULT_MAX_WAIT_MS, 100L, () -> "Failed to consume records for share consumer");
-
-            TestUtils.waitForCondition(() -> {
-                try {
-                    shareConsumer.poll(Duration.ofMillis(500));
-                } catch (Exception e) {
-                    throw new NoRetryException(e);
-                }
-                return callbackCalled.get();
-            }, DEFAULT_MAX_WAIT_MS, 100L, () -> "Received unexpected exception or callback not called");
-            verifyShareGroupStateTopicRecordsProduced();
-        }
-    }
-
-    private static class TestableAcknowledgementCommitCallbackThrows implements AcknowledgementCommitCallback {
-        private final AtomicBoolean callbackCalled;
-
-        public TestableAcknowledgementCommitCallbackThrows(AtomicBoolean callbackCalled) {
-            this.callbackCalled = callbackCalled;
-        }
-
-        @Override
-        public void onComplete(Map<TopicIdPartition, Set<Long>> offsetsMap, Exception exception) {
-            callbackCalled.set(true);
-            throw new org.apache.kafka.common.errors.OutOfOrderSequenceException("Exception thrown in TestableAcknowledgementCommitCallbackThrows.onComplete");
-        }
-    }
-
-    /**
      * Test to verify that calling Thread.interrupt() before ShareConsumer.poll(Duration)
      * causes it to throw InterruptException
      */
@@ -1968,10 +1571,10 @@ public class ShareConsumerTest {
 
     @ClusterTests({
         @ClusterTest(serverProperties = {
-            @ClusterConfigProperty(key = GroupCoordinatorConfig.CONSUMER_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "0")
+            @ClusterConfigProperty(key = GroupCoordinatorConfig.SHARE_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "0")
         }),
         @ClusterTest(serverProperties = {
-            @ClusterConfigProperty(key = GroupCoordinatorConfig.CONSUMER_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "1000")
+            @ClusterConfigProperty(key = GroupCoordinatorConfig.SHARE_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "1000")
         })
     })
     public void testSubscriptionFollowedByTopicCreation() throws InterruptedException {
@@ -2004,10 +1607,10 @@ public class ShareConsumerTest {
 
     @ClusterTests({
         @ClusterTest(serverProperties = {
-            @ClusterConfigProperty(key = GroupCoordinatorConfig.CONSUMER_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "0")
+            @ClusterConfigProperty(key = GroupCoordinatorConfig.SHARE_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "0")
         }),
         @ClusterTest(serverProperties = {
-            @ClusterConfigProperty(key = GroupCoordinatorConfig.CONSUMER_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "1000")
+            @ClusterConfigProperty(key = GroupCoordinatorConfig.SHARE_GROUP_ASSIGNMENT_INTERVAL_MS_CONFIG, value = "1000")
         })
     })
     public void testSubscriptionAndPollFollowedByTopicDeletion() throws InterruptedException, ExecutionException {
@@ -2093,193 +1696,6 @@ public class ShareConsumerTest {
             messageCount = consumeMessages(new AtomicInteger(0), 0, groupId, 1, 5, true);
             assertEquals(0, messageCount);
             verifyShareGroupStateTopicRecordsProduced();
-        }
-    }
-
-    @ClusterTest
-    public void testShareAutoOffsetResetDefaultValue() {
-        try (ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer("group1");
-             Producer<byte[], byte[]> producer = createProducer()) {
-
-            shareConsumer.subscribe(Set.of(tp.topic()));
-            ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(), null, "key".getBytes(), "value".getBytes());
-            // Producing a record.
-            producer.send(record);
-            producer.flush();
-            ConsumerRecords<byte[], byte[]> records = waitedPoll(shareConsumer, 2500L, 0, true, "group1", List.of(tp));
-            // No records should be consumed because share.auto.offset.reset has a default of "latest". Since the record
-            // was produced before share partition was initialized (which happens after the first share fetch request
-            // in the poll method), the start offset would be the latest offset, i.e. 1 (the next offset after the already
-            // present 0th record)
-            assertEquals(0, records.count());
-            // Producing another record.
-            producer.send(record);
-            producer.flush();
-            records = shareConsumer.poll(Duration.ofMillis(5000));
-            // Now the next record should be consumed successfully
-            assertEquals(1, records.count());
-            verifyShareGroupStateTopicRecordsProduced();
-        }
-    }
-
-    @ClusterTest
-    public void testShareAutoOffsetResetEarliest() {
-        alterShareAutoOffsetReset("group1", "earliest");
-        try (ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer("group1");
-             Producer<byte[], byte[]> producer = createProducer()) {
-
-            shareConsumer.subscribe(Set.of(tp.topic()));
-            ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(), null, "key".getBytes(), "value".getBytes());
-            // Producing a record.
-            producer.send(record);
-            producer.flush();
-            ConsumerRecords<byte[], byte[]> records = waitedPoll(shareConsumer, 2500L, 1);
-            // Since the value for share.auto.offset.reset has been altered to "earliest", the consumer should consume
-            // all messages present on the partition
-            assertEquals(1, records.count());
-            // Producing another record.
-            producer.send(record);
-            producer.flush();
-            records = shareConsumer.poll(Duration.ofMillis(5000));
-            // The next records should also be consumed successfully
-            assertEquals(1, records.count());
-            verifyShareGroupStateTopicRecordsProduced();
-        }
-    }
-
-    @ClusterTest
-    public void testShareAutoOffsetResetEarliestAfterLsoMovement() {
-        alterShareAutoOffsetReset("group1", "earliest");
-        try (
-            Producer<byte[], byte[]> producer = createProducer();
-            Admin adminClient = createAdminClient()
-        ) {
-
-            ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(), null, "key".getBytes(), "value".getBytes());
-            // We write 10 records to the topic, so they would be written from offsets 0-9 on the topic.
-            for (int i = 0; i < 10; i++) {
-                assertDoesNotThrow(() -> producer.send(record).get(), "Failed to send records");
-            }
-
-            // We delete records before offset 5, so the LSO should move to 5.
-            assertDoesNotThrow(() -> adminClient.deleteRecords(Map.of(tp, RecordsToDelete.beforeOffset(5L))).all().get(), "Failed to delete records");
-
-            int consumedMessageCount = consumeMessages(new AtomicInteger(0), 5, "group1", 1, 10, true);
-            // The records returned belong to offsets 5-9.
-            assertEquals(5, consumedMessageCount);
-            verifyShareGroupStateTopicRecordsProduced();
-        }
-    }
-
-    @ClusterTest
-    public void testShareAutoOffsetResetMultipleGroupsWithDifferentValue() {
-        alterShareAutoOffsetReset("group1", "earliest");
-        alterShareAutoOffsetReset("group2", "latest");
-        try (ShareConsumer<byte[], byte[]> shareConsumerEarliest = createShareConsumer("group1");
-             ShareConsumer<byte[], byte[]> shareConsumerLatest = createShareConsumer("group2");
-             Producer<byte[], byte[]> producer = createProducer()) {
-
-            shareConsumerEarliest.subscribe(Set.of(tp.topic()));
-            shareConsumerLatest.subscribe(Set.of(tp.topic()));
-
-            ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(), null, "key".getBytes(), "value".getBytes());
-            // Producing a record.
-            producer.send(record);
-            producer.flush();
-            ConsumerRecords<byte[], byte[]> records1 = waitedPoll(shareConsumerEarliest, 2500L, 1);
-            // Since the value for share.auto.offset.reset has been altered to "earliest", the consumer should consume
-            // all messages present on the partition
-            assertEquals(1, records1.count());
-
-            ConsumerRecords<byte[], byte[]> records2 = waitedPoll(shareConsumerLatest, 2500L, 0, true, "group2", List.of(tp));
-            // Since the value for share.auto.offset.reset has been altered to "latest", the consumer should not consume
-            // any message
-            assertEquals(0, records2.count());
-
-            // Producing another record.
-            producer.send(record);
-
-            records1 = shareConsumerEarliest.poll(Duration.ofMillis(5000));
-            // The next record should also be consumed successfully by group1
-            assertEquals(1, records1.count());
-
-            records2 = shareConsumerLatest.poll(Duration.ofMillis(5000));
-            // The next record should also be consumed successfully by group2
-            assertEquals(1, records2.count());
-            verifyShareGroupStateTopicRecordsProduced();
-        }
-    }
-
-    @ClusterTest
-    public void testShareAutoOffsetResetByDuration() throws Exception {
-        // Set auto offset reset to 1 hour before current time
-        alterShareAutoOffsetReset("group1", "by_duration:PT1H");
-
-        try (ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer("group1");
-             Producer<byte[], byte[]> producer = createProducer()) {
-
-            long currentTime = System.currentTimeMillis();
-            long twoHoursAgo = currentTime - TimeUnit.HOURS.toMillis(2);
-            long thirtyMinsAgo = currentTime - TimeUnit.MINUTES.toMillis(30);
-
-            // Produce messages with different timestamps
-            ProducerRecord<byte[], byte[]> oldRecord = new ProducerRecord<>(tp.topic(), tp.partition(),
-                twoHoursAgo, "old_key".getBytes(), "old_value".getBytes());
-            ProducerRecord<byte[], byte[]> recentRecord = new ProducerRecord<>(tp.topic(), tp.partition(),
-                thirtyMinsAgo, "recent_key".getBytes(), "recent_value".getBytes());
-            ProducerRecord<byte[], byte[]> currentRecord = new ProducerRecord<>(tp.topic(), tp.partition(),
-                currentTime, "current_key".getBytes(), "current_value".getBytes());
-
-            producer.send(oldRecord).get();
-            producer.send(recentRecord).get();
-            producer.send(currentRecord).get();
-            producer.flush();
-
-            shareConsumer.subscribe(Set.of(tp.topic()));
-
-            // Should only receive messages from last hour (recent and current)
-            List<ConsumerRecord<byte[], byte[]>> records = consumeRecords(shareConsumer, 2);
-            assertEquals(2, records.size());
-
-            // Verify timestamps and order
-            assertEquals(thirtyMinsAgo, records.get(0).timestamp());
-            assertEquals("recent_key", new String(records.get(0).key()));
-            assertEquals(currentTime, records.get(1).timestamp());
-            assertEquals("current_key", new String(records.get(1).key()));
-        }
-
-        // Set the auto offset reset to 3 hours before current time
-        // so the consumer should consume all messages (3 records)
-        alterShareAutoOffsetReset("group2", "by_duration:PT3H");
-        try (ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer("group2")) {
-
-            shareConsumer.subscribe(Set.of(tp.topic()));
-            List<ConsumerRecord<byte[], byte[]>> records = consumeRecords(shareConsumer, 3);
-            assertEquals(3, records.size());
-            verifyShareGroupStateTopicRecordsProduced();
-        }
-    }
-
-    @ClusterTest
-    public void testShareAutoOffsetResetByDurationInvalidFormat() {
-        // Test invalid duration format
-        ConfigResource configResource = new ConfigResource(ConfigResource.Type.GROUP, "group1");
-        Map<ConfigResource, Collection<AlterConfigOp>> alterEntries = new HashMap<>();
-
-        // Test invalid duration format
-        try (Admin adminClient = createAdminClient()) {
-            alterEntries.put(configResource, List.of(new AlterConfigOp(new ConfigEntry(
-                GroupConfig.SHARE_AUTO_OFFSET_RESET_CONFIG, "by_duration:1h"), AlterConfigOp.OpType.SET)));
-            ExecutionException e1 = assertThrows(ExecutionException.class, () ->
-                adminClient.incrementalAlterConfigs(alterEntries).all().get());
-            assertInstanceOf(InvalidConfigurationException.class, e1.getCause());
-
-            // Test negative duration
-            alterEntries.put(configResource, List.of(new AlterConfigOp(new ConfigEntry(
-                GroupConfig.SHARE_AUTO_OFFSET_RESET_CONFIG, "by_duration:-PT1H"), AlterConfigOp.OpType.SET)));
-            ExecutionException e2 = assertThrows(ExecutionException.class, () ->
-                adminClient.incrementalAlterConfigs(alterEntries).all().get());
-            assertInstanceOf(InvalidConfigurationException.class, e2.getCause());
         }
     }
 
@@ -2420,80 +1836,6 @@ public class ShareConsumerTest {
         shutdownExecutorService(service);
 
         verifyShareGroupStateTopicRecordsProduced();
-    }
-
-    @ClusterTest
-    public void testDeliveryCountDifferentBehaviorWhenClosingSessionWithExplicitAcknowledgement() {
-        alterShareAutoOffsetReset("group1", "earliest");
-        try (Producer<byte[], byte[]> producer = createProducer();
-            ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer(
-                "group1",
-                Map.of(ConsumerConfig.SHARE_ACKNOWLEDGEMENT_MODE_CONFIG, EXPLICIT))) {
-
-            ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(), null,
-                "key".getBytes(), "value".getBytes());
-            producer.send(record);
-            producer.send(record);
-            producer.flush();
-
-            shareConsumer.subscribe(Set.of(tp.topic()));
-            ConsumerRecords<byte[], byte[]> records = waitedPoll(shareConsumer, 2500L, 2);
-            assertEquals(2, records.count());
-            // Acknowledge the first record with AcknowledgeType.RELEASE
-            shareConsumer.acknowledge(records.records(tp).get(0), AcknowledgeType.RELEASE);
-            Map<TopicIdPartition, Optional<KafkaException>> result = shareConsumer.commitSync();
-            assertEquals(1, result.size());
-        }
-
-        // Test delivery count
-        try (ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer("group1", Map.of())) {
-            shareConsumer.subscribe(Set.of(tp.topic()));
-            ConsumerRecords<byte[], byte[]> records = waitedPoll(shareConsumer, 2500L, 2);
-            assertEquals(2, records.count());
-            assertEquals((short) 2, records.records(tp).get(0).deliveryCount().get());
-            assertEquals((short) 2, records.records(tp).get(1).deliveryCount().get());
-        }
-    }
-
-    @ClusterTest(
-        serverProperties = {
-            @ClusterConfigProperty(key = "group.share.delivery.count.limit", value = "3"),
-        }
-    )
-    public void testBehaviorOnDeliveryCountBoundary() {
-        alterShareAutoOffsetReset("group1", "earliest");
-        try (Producer<byte[], byte[]> producer = createProducer();
-            ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer(
-                "group1",
-                Map.of(ConsumerConfig.SHARE_ACKNOWLEDGEMENT_MODE_CONFIG, EXPLICIT))) {
-
-            ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(), null,
-                "key".getBytes(), "value".getBytes());
-            producer.send(record);
-            producer.flush();
-
-            shareConsumer.subscribe(Set.of(tp.topic()));
-            ConsumerRecords<byte[], byte[]> records = waitedPoll(shareConsumer, 2500L, 1);
-            assertEquals(1, records.count());
-            assertEquals((short) 1, records.records(tp).get(0).deliveryCount().get());
-            // Acknowledge the record with AcknowledgeType.RELEASE.
-            shareConsumer.acknowledge(records.records(tp).get(0), AcknowledgeType.RELEASE);
-            Map<TopicIdPartition, Optional<KafkaException>> result = shareConsumer.commitSync();
-            assertEquals(1, result.size());
-
-            // Consume again, the delivery count should be 2.
-            records = waitedPoll(shareConsumer, 2500L, 1);
-            assertEquals(1, records.count());
-            assertEquals((short) 2, records.records(tp).get(0).deliveryCount().get());
-        }
-
-        // Start again and same record should be delivered
-        try (ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer("group1", Map.of())) {
-            shareConsumer.subscribe(Set.of(tp.topic()));
-            ConsumerRecords<byte[], byte[]> records = waitedPoll(shareConsumer, 2500L, 1);
-            assertEquals(1, records.count());
-            assertEquals((short) 3, records.records(tp).get(0).deliveryCount().get());
-        }
     }
 
     @ClusterTest(
@@ -2646,487 +1988,6 @@ public class ShareConsumerTest {
     }
 
     @ClusterTest
-    public void testReadCommittedIsolationLevel() {
-        alterShareAutoOffsetReset("group1", "earliest");
-        alterShareIsolationLevel("group1", "read_committed");
-        try (Producer<byte[], byte[]> transactionalProducer = createProducer("T1");
-             ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer("group1")) {
-            produceCommittedAndAbortedTransactionsInInterval(transactionalProducer, 10, 5);
-            shareConsumer.subscribe(Set.of(tp.topic()));
-            ConsumerRecords<byte[], byte[]> records = waitedPoll(shareConsumer, 5000L, 8);
-            // 5th and 10th message transaction was aborted, hence they won't be included in the fetched records.
-            assertEquals(8, records.count());
-            int messageCounter = 1;
-            for (ConsumerRecord<byte[], byte[]> record : records) {
-                assertEquals(tp.topic(), record.topic());
-                assertEquals(tp.partition(), record.partition());
-                if (messageCounter % 5 == 0)
-                    messageCounter++;
-                assertEquals("Message " + messageCounter, new String(record.value()));
-                messageCounter++;
-            }
-        }
-        verifyShareGroupStateTopicRecordsProduced();
-    }
-
-    @ClusterTest
-    public void testReadUncommittedIsolationLevel() {
-        alterShareAutoOffsetReset("group1", "earliest");
-        alterShareIsolationLevel("group1", "read_uncommitted");
-        try (Producer<byte[], byte[]> transactionalProducer = createProducer("T1");
-             ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer("group1")) {
-            produceCommittedAndAbortedTransactionsInInterval(transactionalProducer, 10, 5);
-            shareConsumer.subscribe(Set.of(tp.topic()));
-            ConsumerRecords<byte[], byte[]> records = waitedPoll(shareConsumer, 5000L, 10);
-            // Even though 5th and 10th message transaction was aborted, they will be included in the fetched records since IsolationLevel is READ_UNCOMMITTED.
-            assertEquals(10, records.count());
-            int messageCounter = 1;
-            for (ConsumerRecord<byte[], byte[]> record : records) {
-                assertEquals(tp.topic(), record.topic());
-                assertEquals(tp.partition(), record.partition());
-                assertEquals("Message " + messageCounter, new String(record.value()));
-                messageCounter++;
-            }
-        }
-        verifyShareGroupStateTopicRecordsProduced();
-    }
-
-    @ClusterTest
-    public void testAlterReadUncommittedToReadCommittedIsolationLevel() {
-        alterShareAutoOffsetReset("group1", "earliest");
-        alterShareIsolationLevel("group1", "read_uncommitted");
-        try (Producer<byte[], byte[]> transactionalProducer = createProducer("T1");
-             ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer("group1", Map.of(
-                 ConsumerConfig.SHARE_ACKNOWLEDGEMENT_MODE_CONFIG, EXPLICIT))) {
-            shareConsumer.subscribe(Set.of(tp.topic()));
-            transactionalProducer.initTransactions();
-            try {
-                // First transaction is committed.
-                produceCommittedTransaction(transactionalProducer, "Message 1");
-
-                ConsumerRecords<byte[], byte[]> records = waitedPoll(shareConsumer, 5000L, 1);
-                assertEquals(1, records.count());
-                ConsumerRecord<byte[], byte[]> record = records.iterator().next();
-                assertEquals("Message 1", new String(record.value()));
-                assertEquals(tp.topic(), record.topic());
-                assertEquals(tp.partition(), record.partition());
-                records.forEach(consumedRecord -> shareConsumer.acknowledge(consumedRecord, AcknowledgeType.ACCEPT));
-                shareConsumer.commitSync();
-
-                // Second transaction is aborted.
-                produceAbortedTransaction(transactionalProducer, "Message 2");
-
-                records = waitedPoll(shareConsumer, 5000L, 1);
-                assertEquals(1, records.count());
-                record = records.iterator().next();
-                assertEquals("Message 2", new String(record.value()));
-                records.forEach(consumedRecord -> shareConsumer.acknowledge(consumedRecord, AcknowledgeType.ACCEPT));
-                shareConsumer.commitSync();
-
-                // Third transaction is committed.
-                produceCommittedTransaction(transactionalProducer, "Message 3");
-                // Fourth transaction is aborted.
-                produceAbortedTransaction(transactionalProducer, "Message 4");
-
-                records = waitedPoll(shareConsumer, 5000L, 2);
-                // Message 3 and Message 4 would be returned by this poll.
-                assertEquals(2, records.count());
-                Iterator<ConsumerRecord<byte[], byte[]>> recordIterator = records.iterator();
-                record = recordIterator.next();
-                assertEquals("Message 3", new String(record.value()));
-                record = recordIterator.next();
-                assertEquals("Message 4", new String(record.value()));
-                // We will make Message 3 and Message 4 available for re-consumption.
-                records.forEach(consumedRecord -> shareConsumer.acknowledge(consumedRecord, AcknowledgeType.RELEASE));
-                shareConsumer.commitSync();
-
-                // We are altering IsolationLevel to READ_COMMITTED now. We will only read committed transactions now.
-                alterShareIsolationLevel("group1", "read_committed");
-
-                // Fifth transaction is committed.
-                produceCommittedTransaction(transactionalProducer, "Message 5");
-                // Sixth transaction is aborted.
-                produceAbortedTransaction(transactionalProducer, "Message 6");
-                // Seventh transaction is aborted.
-                produceAbortedTransaction(transactionalProducer, "Message 7");
-                // Eighth transaction is committed.
-                produceCommittedTransaction(transactionalProducer, "Message 8");
-
-                // Since isolation level is READ_COMMITTED, we can consume Message 3 (committed transaction that was released), Message 5 and Message 8.
-                // We cannot consume Message 4 (aborted transaction that was released), Message 6 and Message 7 since they were aborted.
-                List<String> messages = new ArrayList<>();
-                TestUtils.waitForCondition(() -> {
-                    ConsumerRecords<byte[], byte[]> pollRecords = shareConsumer.poll(Duration.ofMillis(5000));
-                    if (pollRecords.count() > 0) {
-                        for (ConsumerRecord<byte[], byte[]> pollRecord : pollRecords)
-                            messages.add(new String(pollRecord.value()));
-                        pollRecords.forEach(consumedRecord -> shareConsumer.acknowledge(consumedRecord, AcknowledgeType.ACCEPT));
-                        shareConsumer.commitSync();
-                    }
-                    return messages.size() == 3;
-                }, DEFAULT_MAX_WAIT_MS, 100L, () -> "Failed to consume all records post altering share isolation level");
-
-                assertEquals("Message 3", messages.get(0));
-                assertEquals("Message 5", messages.get(1));
-                assertEquals("Message 8", messages.get(2));
-            } catch (InterruptedException e) {
-                throw new RuntimeException(e);
-            } finally {
-                transactionalProducer.close();
-            }
-        }
-        verifyShareGroupStateTopicRecordsProduced();
-    }
-
-    @ClusterTest
-    public void testAlterReadCommittedToReadUncommittedIsolationLevelWithReleaseAck() {
-        alterShareAutoOffsetReset("group1", "earliest");
-        alterShareIsolationLevel("group1", "read_committed");
-        try (Producer<byte[], byte[]> transactionalProducer = createProducer("T1");
-             ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer("group1", Map.of(
-                 ConsumerConfig.SHARE_ACKNOWLEDGEMENT_MODE_CONFIG, EXPLICIT))) {
-            shareConsumer.subscribe(Set.of(tp.topic()));
-            transactionalProducer.initTransactions();
-
-            try {
-                // First transaction is committed.
-                produceCommittedTransaction(transactionalProducer, "Message 1");
-
-                ConsumerRecords<byte[], byte[]> records = waitedPoll(shareConsumer, 5000L, 1);
-                assertEquals(1, records.count());
-                ConsumerRecord<byte[], byte[]> record = records.iterator().next();
-                assertEquals("Message 1", new String(record.value()));
-                assertEquals(tp.topic(), record.topic());
-                assertEquals(tp.partition(), record.partition());
-                records.forEach(consumedRecord -> shareConsumer.acknowledge(consumedRecord, AcknowledgeType.ACCEPT));
-                shareConsumer.commitSync();
-
-                // Second transaction is aborted.
-                produceAbortedTransaction(transactionalProducer, "Message 2");
-
-                // Setting the acknowledgement commit callback to verify acknowledgement completion.
-                Map<TopicPartition, Set<Long>> partitionOffsetsMap = new HashMap<>();
-                shareConsumer.setAcknowledgementCommitCallback(new TestableAcknowledgementCommitCallback(partitionOffsetsMap, Map.of()));
-
-                // We will not receive any records since the transaction for Message 2 was aborted. Wait for the
-                // aborted marker offset for Message 2 (3L) to be fetched and acknowledged by the consumer.
-                TestUtils.waitForCondition(() -> {
-                    ConsumerRecords<byte[], byte[]> pollRecords = shareConsumer.poll(Duration.ofMillis(500));
-                    return pollRecords.count() == 0 && partitionOffsetsMap.containsKey(tp) && partitionOffsetsMap.get(tp).contains(3L);
-                }, DEFAULT_MAX_WAIT_MS, 100L, () -> "Failed to consume abort transaction and marker offset for Message 2");
-
-                // Third transaction is committed.
-                produceCommittedTransaction(transactionalProducer, "Message 3");
-                // Fourth transaction is aborted.
-                produceAbortedTransaction(transactionalProducer, "Message 4");
-
-                // Setting the acknowledgement commit callback to verify acknowledgement completion.
-                Map<TopicPartition, Set<Long>> partitionOffsetsMap2 = new HashMap<>();
-                shareConsumer.setAcknowledgementCommitCallback(new TestableAcknowledgementCommitCallback(partitionOffsetsMap2, Map.of()));
-
-                records = waitedPoll(shareConsumer, 5000L, 1);
-                // Message 3 would be returned by this poll.
-                assertEquals(1, records.count());
-                record = records.iterator().next();
-                assertEquals("Message 3", new String(record.value()));
-                // We will make Message 3 available for re-consumption.
-                records.forEach(consumedRecord -> shareConsumer.acknowledge(consumedRecord, AcknowledgeType.RELEASE));
-                shareConsumer.commitSync();
-
-                // Wait for the aborted marker offset for Message 4 (7L) to be fetched and acknowledged by the consumer.
-                TestUtils.waitForCondition(() -> {
-                    ConsumerRecords<byte[], byte[]> pollRecords = shareConsumer.poll(Duration.ofMillis(500));
-                    if (pollRecords.count() > 0) {
-                        // We will release Message 3 again if it was received in this poll().
-                        pollRecords.forEach(consumerRecord -> shareConsumer.acknowledge(consumerRecord, AcknowledgeType.RELEASE));
-                    }
-                    return partitionOffsetsMap2.containsKey(tp) && partitionOffsetsMap2.get(tp).contains(7L);
-                }, DEFAULT_MAX_WAIT_MS, 100L, () -> "Failed to consume abort transaction marker offset for Message 4");
-
-                // We are altering IsolationLevel to READ_UNCOMMITTED now. We will read both committed/aborted transactions now.
-                alterShareIsolationLevel("group1", "read_uncommitted");
-
-                // Fifth transaction is committed.
-                produceCommittedTransaction(transactionalProducer, "Message 5");
-                // Sixth transaction is aborted.
-                produceAbortedTransaction(transactionalProducer, "Message 6");
-                // Seventh transaction is aborted.
-                produceAbortedTransaction(transactionalProducer, "Message 7");
-                // Eighth transaction is committed.
-                produceCommittedTransaction(transactionalProducer, "Message 8");
-
-                // Since isolation level is READ_UNCOMMITTED, we can consume Message 3 (committed transaction that was released), Message 5, Message 6, Message 7 and Message 8.
-                Set<String> finalMessages = new HashSet<>();
-                TestUtils.waitForCondition(() -> {
-                    ConsumerRecords<byte[], byte[]> pollRecords = shareConsumer.poll(Duration.ofMillis(5000));
-                    if (pollRecords.count() > 0) {
-                        for (ConsumerRecord<byte[], byte[]> pollRecord : pollRecords)
-                            finalMessages.add(new String(pollRecord.value()));
-                        pollRecords.forEach(consumedRecord -> shareConsumer.acknowledge(consumedRecord, AcknowledgeType.ACCEPT));
-                        shareConsumer.commitSync();
-                    }
-                    return finalMessages.size() == 5;
-                }, DEFAULT_MAX_WAIT_MS, 100L, () -> "Failed to consume all records post altering share isolation level");
-
-                Set<String> expected = Set.of("Message 3", "Message 5", "Message 6", "Message 7", "Message 8");
-                assertEquals(expected, finalMessages);
-            } catch (InterruptedException e) {
-                throw new RuntimeException(e);
-            } finally {
-                transactionalProducer.close();
-            }
-        }
-        verifyShareGroupStateTopicRecordsProduced();
-    }
-
-    @ClusterTest
-    public void testAlterReadCommittedToReadUncommittedIsolationLevelWithRejectAck() {
-        alterShareAutoOffsetReset("group1", "earliest");
-        alterShareIsolationLevel("group1", "read_committed");
-        try (Producer<byte[], byte[]> transactionalProducer = createProducer("T1");
-             ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer("group1", Map.of(
-                 ConsumerConfig.SHARE_ACKNOWLEDGEMENT_MODE_CONFIG, EXPLICIT))) {
-            shareConsumer.subscribe(Set.of(tp.topic()));
-            transactionalProducer.initTransactions();
-
-            try {
-                // First transaction is committed.
-                produceCommittedTransaction(transactionalProducer, "Message 1");
-
-                ConsumerRecords<byte[], byte[]> records = waitedPoll(shareConsumer, 5000L, 1);
-                assertEquals(1, records.count());
-                ConsumerRecord<byte[], byte[]> record = records.iterator().next();
-                assertEquals("Message 1", new String(record.value()));
-                assertEquals(tp.topic(), record.topic());
-                assertEquals(tp.partition(), record.partition());
-                records.forEach(consumedRecord -> shareConsumer.acknowledge(consumedRecord, AcknowledgeType.ACCEPT));
-                shareConsumer.commitSync();
-
-                // Second transaction is aborted.
-                produceAbortedTransaction(transactionalProducer, "Message 2");
-
-                // Setting the acknowledgement commit callback to verify acknowledgement completion.
-                Map<TopicPartition, Set<Long>> partitionOffsetsMap = new HashMap<>();
-                shareConsumer.setAcknowledgementCommitCallback(new TestableAcknowledgementCommitCallback(partitionOffsetsMap, Map.of()));
-
-                // We will not receive any records since the transaction for Message 2 was aborted. Wait for the
-                // aborted marker offset for Message 2 (3L) to be fetched and acknowledged by the consumer.
-                TestUtils.waitForCondition(() -> {
-                    ConsumerRecords<byte[], byte[]> pollRecords = shareConsumer.poll(Duration.ofMillis(500));
-                    return pollRecords.count() == 0 && partitionOffsetsMap.containsKey(tp) && partitionOffsetsMap.get(tp).contains(3L);
-                }, DEFAULT_MAX_WAIT_MS, 100L, () -> "Failed to consume abort transaction and marker offset for Message 2");
-
-                // Third transaction is committed.
-                produceCommittedTransaction(transactionalProducer, "Message 3");
-                // Fourth transaction is aborted.
-                produceAbortedTransaction(transactionalProducer, "Message 4");
-
-                // Setting the acknowledgement commit callback to verify acknowledgement completion.
-                Map<TopicPartition, Set<Long>> partitionOffsetsMap2 = new HashMap<>();
-                shareConsumer.setAcknowledgementCommitCallback(new TestableAcknowledgementCommitCallback(partitionOffsetsMap2, Map.of()));
-
-                records = waitedPoll(shareConsumer, 5000L, 1);
-                // Message 3 would be returned by this poll.
-                assertEquals(1, records.count());
-                record = records.iterator().next();
-                assertEquals("Message 3", new String(record.value()));
-                // We will make Message 3 available for re-consumption.
-                records.forEach(consumedRecord -> shareConsumer.acknowledge(consumedRecord, AcknowledgeType.REJECT));
-                shareConsumer.commitSync();
-
-                // Wait for the aborted marker offset for Message 4 (7L) to be fetched and acknowledged by the consumer.
-                TestUtils.waitForCondition(() -> {
-                    shareConsumer.poll(Duration.ofMillis(500));
-                    return partitionOffsetsMap2.containsKey(tp) && partitionOffsetsMap2.get(tp).contains(7L);
-                }, DEFAULT_MAX_WAIT_MS, 100L, () -> "Failed to consume abort transaction marker offset for Message 4");
-
-                // We are altering IsolationLevel to READ_UNCOMMITTED now. We will read both committed/aborted transactions now.
-                alterShareIsolationLevel("group1", "read_uncommitted");
-
-                // Fifth transaction is committed.
-                produceCommittedTransaction(transactionalProducer, "Message 5");
-                // Sixth transaction is aborted.
-                produceAbortedTransaction(transactionalProducer, "Message 6");
-                // Seventh transaction is aborted.
-                produceAbortedTransaction(transactionalProducer, "Message 7");
-                // Eighth transaction is committed.
-                produceCommittedTransaction(transactionalProducer, "Message 8");
-
-                // Since isolation level is READ_UNCOMMITTED, we can consume Message 5, Message 6, Message 7 and Message 8.
-                List<String> finalMessages = new ArrayList<>();
-                TestUtils.waitForCondition(() -> {
-                    ConsumerRecords<byte[], byte[]> pollRecords = shareConsumer.poll(Duration.ofMillis(5000));
-                    if (pollRecords.count() > 0) {
-                        for (ConsumerRecord<byte[], byte[]> pollRecord : pollRecords)
-                            finalMessages.add(new String(pollRecord.value()));
-                        pollRecords.forEach(consumedRecord -> shareConsumer.acknowledge(consumedRecord, AcknowledgeType.ACCEPT));
-                        shareConsumer.commitSync();
-                    }
-                    return finalMessages.size() == 4;
-                }, DEFAULT_MAX_WAIT_MS, 100L, () -> "Failed to consume all records post altering share isolation level");
-
-                assertEquals("Message 5", finalMessages.get(0));
-                assertEquals("Message 6", finalMessages.get(1));
-                assertEquals("Message 7", finalMessages.get(2));
-                assertEquals("Message 8", finalMessages.get(3));
-            } catch (InterruptedException e) {
-                throw new RuntimeException(e);
-            } finally {
-                transactionalProducer.close();
-            }
-        }
-        verifyShareGroupStateTopicRecordsProduced();
-    }
-
-    @ClusterTest
-    public void testPollInBatchOptimizedMode() {
-        alterShareAutoOffsetReset("group1", "earliest");
-        try (Producer<byte[], byte[]> producer = createProducer();
-             ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer(
-                "group1",
-                Map.of(ConsumerConfig.MAX_POLL_RECORDS_CONFIG, 5, ConsumerConfig.SHARE_ACQUIRE_MODE_CONFIG, BATCH_OPTIMIZED))
-        ) {
-            ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(), null, "key".getBytes(), "value".getBytes());
-            for (int i = 0; i < 10; i++) {
-                producer.send(record);
-            }
-            producer.flush();
-            shareConsumer.subscribe(List.of(tp.topic()));
-
-            // although max.poll.records is set to 5, in batch optimized mode we will still get all 10 records.
-            ConsumerRecords<byte[], byte[]> records = waitedPoll(shareConsumer, 2500L, 10);
-            assertEquals(10, records.count());
-            verifyShareGroupStateTopicRecordsProduced();
-        }
-    }
-
-    @ClusterTest
-    public void testPollInRecordLimitMode() {
-        alterShareAutoOffsetReset("group1", "earliest");
-        try (Producer<byte[], byte[]> producer = createProducer();
-             ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer(
-                 "group1",
-                 Map.of(ConsumerConfig.MAX_POLL_RECORDS_CONFIG, 5, ConsumerConfig.SHARE_ACQUIRE_MODE_CONFIG, RECORD_LIMIT))
-        ) {
-            ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(), null, "key".getBytes(), "value".getBytes());
-            for (int i = 0; i < 10; i++) {
-                producer.send(record);
-            }
-            producer.flush();
-            shareConsumer.subscribe(List.of(tp.topic()));
-
-            // In record limit mode we will get only up to max.poll.records number of records.
-            ConsumerRecords<byte[], byte[]> records = waitedPoll(shareConsumer, 2500L, 5);
-            assertEquals(5, records.count());
-            verifyShareGroupStateTopicRecordsProduced();
-        }
-    }
-
-    @ClusterTest
-    public void testPollAndExplicitAcknowledgeSingleMessageInRecordLimitMode() {
-        alterShareAutoOffsetReset("group1", "earliest");
-        try (Producer<byte[], byte[]> producer = createProducer();
-             ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer(
-                 "group1",
-                 Map.of(ConsumerConfig.MAX_POLL_RECORDS_CONFIG, 1,
-                     ConsumerConfig.SHARE_ACQUIRE_MODE_CONFIG, RECORD_LIMIT,
-                     ConsumerConfig.SHARE_ACKNOWLEDGEMENT_MODE_CONFIG, EXPLICIT))
-        ) {
-            ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(), null, "key".getBytes(), "value".getBytes());
-            for (int i = 0; i < 10; i++) {
-                producer.send(record);
-            }
-            producer.flush();
-            shareConsumer.subscribe(List.of(tp.topic()));
-
-            for (int i = 0; i < 10; i++) {
-                ConsumerRecords<byte[], byte[]> records = waitedPoll(shareConsumer, 2500L, 1);
-                assertEquals(1, records.count());
-                for (ConsumerRecord<byte[], byte[]> rec : records) {
-                    shareConsumer.acknowledge(rec, AcknowledgeType.ACCEPT);
-                }
-                shareConsumer.commitSync();
-            }
-            verifyShareGroupStateTopicRecordsProduced();
-        }
-    }
-
-    @ClusterTest
-    public void testExplicitAcknowledgeSuccessInRecordLimitMode() {
-        alterShareAutoOffsetReset("group1", "earliest");
-        try (Producer<byte[], byte[]> producer = createProducer();
-             ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer(
-                 "group1",
-                 Map.of(ConsumerConfig.MAX_POLL_RECORDS_CONFIG, 10,
-                     ConsumerConfig.SHARE_ACQUIRE_MODE_CONFIG, RECORD_LIMIT,
-                     ConsumerConfig.SHARE_ACKNOWLEDGEMENT_MODE_CONFIG, EXPLICIT))
-        ) {
-            ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(), null, "key".getBytes(), "value".getBytes());
-            for (int i = 0; i < 15; i++) {
-                producer.send(record);
-            }
-            producer.flush();
-            shareConsumer.subscribe(List.of(tp.topic()));
-            ConsumerRecords<byte[], byte[]> records = waitedPoll(shareConsumer, 2500L, 10);
-            assertEquals(10, records.count());
-            for (ConsumerRecord<byte[], byte[]> rec : records) {
-                shareConsumer.acknowledge(rec, AcknowledgeType.ACCEPT);
-            }
-            shareConsumer.commitSync();
-
-            records = waitedPoll(shareConsumer, 2500L, 5);
-            assertEquals(5, records.count());
-            verifyShareGroupStateTopicRecordsProduced();
-        }
-    }
-
-    @ClusterTest
-    public void testExplicitAcknowledgeReleaseAcceptInRecordLimitMode() {
-        alterShareAutoOffsetReset("group1", "earliest");
-        try (Producer<byte[], byte[]> producer = createProducer();
-             ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer(
-                 "group1",
-                 Map.of(ConsumerConfig.MAX_POLL_RECORDS_CONFIG, 10,
-                     ConsumerConfig.SHARE_ACQUIRE_MODE_CONFIG, RECORD_LIMIT,
-                     ConsumerConfig.SHARE_ACKNOWLEDGEMENT_MODE_CONFIG, EXPLICIT))
-        ) {
-            for (int i = 0; i < 20; i++) {
-                ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp2.topic(), tp2.partition(), null, "key".getBytes(), ("Message " + i).getBytes());
-                producer.send(record);
-            }
-            producer.flush();
-            shareConsumer.subscribe(List.of(tp2.topic()));
-            ConsumerRecords<byte[], byte[]> records = waitedPoll(shareConsumer, 2500L, 10);
-            assertEquals(10, records.count());
-
-            int count = 0;
-            Map<TopicIdPartition, Optional<KafkaException>> result;
-            for (ConsumerRecord<byte[], byte[]> record : records) {
-                if (count % 2 == 0) {
-                    shareConsumer.acknowledge(record, AcknowledgeType.ACCEPT);
-                } else {
-                    shareConsumer.acknowledge(record, AcknowledgeType.RELEASE);
-                }
-                result = shareConsumer.commitSync();
-                assertEquals(1, result.size());
-                count++;
-            }
-
-            // Poll again to get 10 records.
-            records = waitedPoll(shareConsumer, 2500L, 10);
-            assertEquals(10, records.count());
-            for (ConsumerRecord<byte[], byte[]> record : records) {
-                shareConsumer.acknowledge(record, AcknowledgeType.ACCEPT);
-            }
-
-            // Get the rest of all 5 records.
-            records = waitedPoll(shareConsumer, 2500L, 5);
-            assertEquals(5, records.count());
-            verifyShareGroupStateTopicRecordsProduced();
-        }
-    }
-
-    @ClusterTest
     public void testCommitSyncFailsForDeletedTopic() throws InterruptedException {
         Uuid topicId = createTopic("baz", 1, 1);
         alterShareAutoOffsetReset("group1", "earliest");
@@ -3155,244 +2016,6 @@ public class ShareConsumerTest {
             assertEquals(1, commitResult.size());
             assertInstanceOf(UnknownTopicIdException.class, commitResult.get(new TopicIdPartition(topicId, 0, "baz")).get());
         }
-    }
-
-    @ClusterTest
-    public void testRenewAcknowledgementOnPoll() {
-        alterShareAutoOffsetReset("group1", "earliest");
-        try (Producer<byte[], byte[]> producer = createProducer();
-             ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer(
-                 "group1",
-                 Map.of(ConsumerConfig.SHARE_ACKNOWLEDGEMENT_MODE_CONFIG, EXPLICIT))
-        ) {
-            AtomicInteger acknowledgementsCommitted = new AtomicInteger(0);
-            shareConsumer.setAcknowledgementCommitCallback((offsetsByTopicPartition, exception) ->
-                offsetsByTopicPartition.forEach((tip, offsets) -> acknowledgementsCommitted.addAndGet(offsets.size())));
-
-            for (int i = 0; i < 10; i++) {
-                ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(), null, "key".getBytes(), ("Message " + i).getBytes());
-                producer.send(record);
-            }
-            producer.flush();
-
-            shareConsumer.subscribe(List.of(tp.topic()));
-            ConsumerRecords<byte[], byte[]> records = waitedPoll(shareConsumer, 2500L, 10);
-            assertEquals(10, records.count());
-
-            int count = 0;
-            for (ConsumerRecord<byte[], byte[]> record : records) {
-                if (count % 2 == 0) {
-                    shareConsumer.acknowledge(record, AcknowledgeType.ACCEPT);
-                } else {
-                    shareConsumer.acknowledge(record, AcknowledgeType.RENEW);
-                }
-                count++;
-            }
-
-            // Get the rest of all 5 records.
-            records = waitedPoll(shareConsumer, 2500L, 5);
-            assertEquals(5, records.count());
-            for (ConsumerRecord<byte[], byte[]> record : records) {
-                shareConsumer.acknowledge(record, AcknowledgeType.ACCEPT);
-            }
-
-            shareConsumer.commitSync();
-            assertEquals(15, acknowledgementsCommitted.get());
-        }
-        verifyYammerMetricCount("ackType=Renew", 5);
-    }
-
-    @ClusterTest
-    public void testRenewAcknowledgementOnCommitSync() {
-        alterShareAutoOffsetReset("group1", "earliest");
-        try (Producer<byte[], byte[]> producer = createProducer();
-             ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer(
-                 "group1",
-                 Map.of(ConsumerConfig.SHARE_ACKNOWLEDGEMENT_MODE_CONFIG, EXPLICIT))
-        ) {
-            AtomicInteger acknowledgementsCommitted = new AtomicInteger(0);
-            shareConsumer.setAcknowledgementCommitCallback((offsetsByTopicPartition, exception) ->
-                offsetsByTopicPartition.forEach((tip, offsets) -> acknowledgementsCommitted.addAndGet(offsets.size())));
-
-            for (int i = 0; i < 10; i++) {
-                ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(), null, "key".getBytes(), ("Message " + i).getBytes());
-                producer.send(record);
-            }
-            producer.flush();
-
-            shareConsumer.subscribe(List.of(tp.topic()));
-            ConsumerRecords<byte[], byte[]> records = waitedPoll(shareConsumer, 2500L, 10);
-            assertEquals(10, records.count());
-            assertEquals(Optional.of(15000), shareConsumer.acquisitionLockTimeoutMs());
-
-            // The updated acquisition lock timeout is only applied when the next poll is called.
-            alterShareRecordLockDurationMs("group1", 25000);
-
-            int count = 0;
-            Map<TopicIdPartition, Optional<KafkaException>> result;
-            for (ConsumerRecord<byte[], byte[]> record : records) {
-                if (count % 2 == 0) {
-                    shareConsumer.acknowledge(record, AcknowledgeType.ACCEPT);
-                } else {
-                    shareConsumer.acknowledge(record, AcknowledgeType.RENEW);
-                }
-                result = shareConsumer.commitSync();
-                assertEquals(1, result.size());
-                assertEquals(Optional.of(15000), shareConsumer.acquisitionLockTimeoutMs());
-                assertEquals(Optional.empty(), result.get(new TopicIdPartition(tpId, tp.partition(), tp.topic())));
-                count++;
-            }
-
-            // Get the rest of all 5 records.
-            records = waitedPoll(shareConsumer, 2500L, 5);
-            assertEquals(5, records.count());
-            assertEquals(Optional.of(25000), shareConsumer.acquisitionLockTimeoutMs());
-            for (ConsumerRecord<byte[], byte[]> record : records) {
-                shareConsumer.acknowledge(record, AcknowledgeType.ACCEPT);
-            }
-        }
-        verifyYammerMetricCount("ackType=Renew", 5);
-    }
-
-    @ClusterTest
-    public void testRenewAcknowledgementInvalidStateRecord() {
-        alterShareAutoOffsetReset("group1", "earliest");
-        try (Producer<byte[], byte[]> producer = createProducer();
-             ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer(
-                 "group1",
-                 Map.of(ConsumerConfig.SHARE_ACKNOWLEDGEMENT_MODE_CONFIG, EXPLICIT))
-        ) {
-            AtomicInteger acknowledgementsCommitted = new AtomicInteger(0);
-            shareConsumer.setAcknowledgementCommitCallback((offsetsByTopicPartition, exception) ->
-                offsetsByTopicPartition.forEach((tip, offsets) -> acknowledgementsCommitted.addAndGet(offsets.size())));
-
-            ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(), null, "key".getBytes(), "Message ".getBytes());
-            producer.send(record);
-            producer.flush();
-
-            shareConsumer.subscribe(List.of(tp.topic()));
-            ConsumerRecords<byte[], byte[]> records = waitedPoll(shareConsumer, 2500L, 1);
-            assertEquals(1, records.count());
-
-            for (ConsumerRecord<byte[], byte[]> rec : records) {
-                shareConsumer.acknowledge(rec, AcknowledgeType.REJECT);
-                shareConsumer.commitSync();
-                assertThrows(IllegalStateException.class, () -> shareConsumer.acknowledge(rec, AcknowledgeType.RENEW));
-            }
-        }
-        verifyYammerMetricCount("ackType=Renew", 0);
-    }
-
-    @ClusterTest
-    public void testRenewAcknowledgementDisabled() {
-        alterShareAutoOffsetReset("group1", "earliest");
-        alterShareRenewAcknowledgeEnable("group1", false);
-        try (Producer<byte[], byte[]> producer = createProducer();
-             ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer(
-                 "group1",
-                 Map.of(ConsumerConfig.SHARE_ACKNOWLEDGEMENT_MODE_CONFIG, EXPLICIT))
-        ) {
-            ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(), null, "key".getBytes(), "Message".getBytes());
-            producer.send(record);
-            producer.flush();
-
-            shareConsumer.subscribe(List.of(tp.topic()));
-            ConsumerRecords<byte[], byte[]> records = waitedPoll(shareConsumer, 2500L, 1);
-            assertEquals(1, records.count());
-
-            for (ConsumerRecord<byte[], byte[]> rec : records) {
-                shareConsumer.acknowledge(rec, AcknowledgeType.RENEW);
-            }
-
-            Map<TopicIdPartition, Optional<KafkaException>> result = shareConsumer.commitSync();
-            assertEquals(1, result.size());
-            Optional<KafkaException> error = result.get(new TopicIdPartition(tpId, tp.partition(), tp.topic()));
-            assertTrue(error.isPresent());
-            assertInstanceOf(InvalidRecordStateException.class, error.get());
-        }
-    }
-
-    @ClusterTest(
-        brokers = 1,
-        serverProperties = {
-            @ClusterConfigProperty(key = "group.share.record.lock.duration.ms", value = "12000"),
-            @ClusterConfigProperty(key = "group.share.min.record.lock.duration.ms", value = "12000"),
-        }
-    )
-    public void testRenewAcknowledgementNoResultInPoll() {
-        alterShareAutoOffsetReset("group1", "earliest");
-        try (Producer<byte[], byte[]> producer = createProducer();
-             ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer(
-                 "group1",
-                 Map.of(ConsumerConfig.SHARE_ACKNOWLEDGEMENT_MODE_CONFIG, EXPLICIT))
-        ) {
-            AtomicInteger acknowledgementsCommitted = new AtomicInteger(0);
-            shareConsumer.setAcknowledgementCommitCallback((offsetsByTopicPartition, exception) ->
-                offsetsByTopicPartition.forEach((tip, offsets) -> acknowledgementsCommitted.addAndGet(offsets.size())));
-
-            for (int i = 0; i < 10; i++) {
-                ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(), null, "key".getBytes(), ("Message " + i).getBytes());
-                producer.send(record);
-            }
-            producer.flush();
-
-            shareConsumer.subscribe(List.of(tp.topic()));
-            ConsumerRecords<byte[], byte[]> records = waitedPoll(shareConsumer, 2500L, 10);
-            assertEquals(10, records.count());
-
-            int count = 0;
-            for (ConsumerRecord<byte[], byte[]> record : records) {
-                if (count % 2 == 0) {
-                    shareConsumer.acknowledge(record, AcknowledgeType.ACCEPT);
-                } else {
-                    shareConsumer.acknowledge(record, AcknowledgeType.RENEW);
-                }
-                count++;
-            }
-
-            // 5 more records (total 15 produced).
-            for (int i = 10; i < 15; i++) {
-                ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(), null, "key".getBytes(), ("Message " + i).getBytes());
-                producer.send(record);
-            }
-
-            // Get the rest of all 5 records.
-            records = waitedPoll(shareConsumer, 11500L, 0);  // This will send the acks but not return next 5 records (10-15)
-            assertEquals(10, acknowledgementsCommitted.get());
-            assertEquals(0, records.count());
-            verifyYammerMetricCount("ackType=Renew", 5);
-
-            // Renewal duration passed, now records will be back.
-            records = waitedPoll(shareConsumer, 2500L, 5);  // Renewed records as well as 10-15 records.
-            assertEquals(5, records.count());
-            for (ConsumerRecord<byte[], byte[]> record : records) {
-                shareConsumer.acknowledge(record, AcknowledgeType.ACCEPT);
-            }
-
-            shareConsumer.commitSync();
-
-            records = waitedPoll(shareConsumer, 2500L, 5);  // 10-15 records.
-            assertEquals(5, records.count());
-            for (ConsumerRecord<byte[], byte[]> record : records) {
-                shareConsumer.acknowledge(record, AcknowledgeType.ACCEPT);
-            }
-
-            shareConsumer.commitSync();
-
-            // Initial - 5 renew + 5 accept, Subsequent - 5 renewed accepted + 5 fresh accepted (10-15)
-            assertEquals(20, acknowledgementsCommitted.get());
-        }
-        verifyYammerMetricCount("ackType=Renew", 5);
-    }
-
-    private void verifyYammerMetricCount(String filterString, int count) {
-        com.yammer.metrics.core.Metric renewAck = KafkaYammerMetrics.defaultRegistry().allMetrics().entrySet().stream()
-            .filter(entry -> entry.getKey().toString().contains(filterString))
-            .findAny()
-            .orElseThrow(() -> new AssertionError("metric not found"))
-            .getValue();
-
-        assertEquals(count, ((Meter) renewAck).count());
     }
 
     @ClusterTest
@@ -3761,1097 +2384,6 @@ public class ShareConsumerTest {
             shareConsumer2.close();
         } catch (InterruptedException | ExecutionException e) {
             fail("Test failed with exception: " + e.getMessage());
-        }
-    }
-
-    @ClusterTest
-    public void testFetchWithThrottledDelivery() {
-        alterShareAutoOffsetReset("group1", "earliest");
-        try (Producer<byte[], byte[]> producer = createProducer();
-            ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer(
-                "group1",
-                Map.of(ConsumerConfig.SHARE_ACKNOWLEDGEMENT_MODE_CONFIG, EXPLICIT))
-        ) {
-            // Produce a batch of 100 messages
-            for (int i = 0; i < 100; i++) {
-                ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(), null, "key".getBytes(), ("Message " + i).getBytes());
-                producer.send(record);
-            }
-            producer.flush();
-
-            shareConsumer.subscribe(List.of(tp.topic()));
-            // Fetch records in 5 iterations, each time acknowledging with RELEASE. 5 is the default
-            // delivery limit hence we should see throttling from Math.ceil(5/2) = 3 fetches.
-            int throttleDeliveryLimit = 3;
-            for (int i = 0; i < 5; i++) {
-                // Adjust expected fetch count based on throttling. If i < throttleDeliveryLimit, we get full batch of 100.
-                // If i == 4 i.e. the last delivery, then we get 1 record.
-                // Otherwise, we get half the previous fetch count due to throttling. In this case, 100 >> (i - throttleDeliveryLimit + 1) it is 50 for i=3.
-                int expectedFetchCount = (i < throttleDeliveryLimit) ? 100 : ((i == 4) ? 1 : 50);
-                ConsumerRecords<byte[], byte[]> records = waitedPoll(shareConsumer, 2500L, expectedFetchCount);
-                assertEquals(expectedFetchCount, records.count());
-
-                records.forEach(record -> shareConsumer.acknowledge(record, AcknowledgeType.RELEASE));
-                Map<TopicIdPartition, Optional<KafkaException>> result = shareConsumer.commitSync();
-                assertEquals(1, result.size());
-                assertEquals(Optional.empty(),
-                    result.get(new TopicIdPartition(tpId, tp.partition(), tp.topic())));
-            }
-
-            // Offset 0 has already reached the delivery limit hence shall be archived.
-            // Offset 1 to 49 shall be in last delivery attempt and hence 1 record per poll.
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 1, 50, 1);
-            // Delivery limit 4.
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 50, 100, 50);
-            // Delivery limit 5.
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 50, 100, 1);
-            // Next poll should not have any records as all records have reached delivery limit.
-            ConsumerRecords<byte[], byte[]> records = shareConsumer.poll(Duration.ofMillis(2500L));
-            assertTrue(records.isEmpty(), "Records should be empty as all records have reached delivery limit. But received: " + records.count());
-        }
-    }
-
-    @ClusterTest(
-        serverProperties = {
-            @ClusterConfigProperty(key = "group.share.delivery.count.limit", value = "10"),
-        }
-    )
-    public void testFetchWithThrottledDeliveryBatchesWithIncreasedDeliveryLimit() {
-        alterShareAutoOffsetReset("group1", "earliest");
-        try (Producer<byte[], byte[]> producer = createProducer();
-            ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer(
-                "group1",
-                Map.of(
-                    ConsumerConfig.SHARE_ACKNOWLEDGEMENT_MODE_CONFIG, EXPLICIT,
-                    ConsumerConfig.MAX_POLL_RECORDS_CONFIG, 512
-                )
-            )
-        ) {
-            // Produce records in complete power of 2 to fully test the throttling behavior.
-            int producedMessageCount = 512;
-            // Produce a batch of 512 messages
-            for (int i = 0; i < producedMessageCount; i++) {
-                ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(),
-                    null, "key".getBytes(), ("Message " + i).getBytes());
-                producer.send(record);
-            }
-            producer.flush();
-
-            // Map which defines expected fetch count for each delivery attempt from 1 to 10.
-            Map<Integer, Integer> expectedFetchCountMap = Map.of(1, 512, 2, 512, 3, 512, 4, 512, 5, 512,
-                6, 256, 7, 128, 8, 64, 9, 32, 10, 1);
-            shareConsumer.subscribe(List.of(tp.topic()));
-            // Fetch records in 10 iterations, each time acknowledging with RELEASE. 10 is the
-            // delivery limit hence we should see throttling from Math.ceil(10/2) = 5 fetches.
-            for (int i = 0; i < 10; i++) {
-                int expectedFetchCount = expectedFetchCountMap.get(i + 1);
-                ConsumerRecords<byte[], byte[]> records = waitedPoll(shareConsumer, 2500L, expectedFetchCount);
-                assertEquals(expectedFetchCount, records.count());
-                // Acknowledge all records with RELEASE.
-                records.forEach(record -> shareConsumer.acknowledge(record, AcknowledgeType.RELEASE));
-                Map<TopicIdPartition, Optional<KafkaException>> result = shareConsumer.commitSync();
-                assertEquals(1, result.size());
-                assertEquals(Optional.empty(), result.get(new TopicIdPartition(tpId, tp.partition(), tp.topic())));
-            }
-
-            // Offset 0 is already verified above, so start from offset 1 and as it's last delivery cycle
-            // hence expectedRecords is 1 for each poll till offset 32.
-            // Delivery 10
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 1, 32, 1);
-            // Delivery 9
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 32, 64, 32);
-            // Delivery 10
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 32, 64, 1);
-            // Delivery 8
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 64, 128, 64);
-            // Delivery 9
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 64, 96, 32);
-            // Delivery 10
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 64, 96, 1);
-            // Delivery 9
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 96, 128, 32);
-            // Delivery 10
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 96, 128, 1);
-            // Delivery 7
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 128, 256, 128);
-            // Delivery 8
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 128, 192, 64);
-            // Delivery 9
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 128, 160, 32);
-            // Delivery 10
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 128, 160, 1);
-            // Delivery 9
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 160, 192, 32);
-            // Delivery 10
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 160, 192, 1);
-            // Delivery 8
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 192, 256, 64);
-            // Delivery 9
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 192, 224, 32);
-            // Delivery 10
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 192, 224, 1);
-            // Delivery 9
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 224, 256, 32);
-            // Delivery 10
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 224, 256, 1);
-            // Delivery 6
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 256, 512, 256);
-            // Delivery 7
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 256, 384, 128);
-            // Delivery 8
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 256, 320, 64);
-            // Delivery 9
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 256, 288, 32);
-            // Delivery 10
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 256, 288, 1);
-            // Delivery 9
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 288, 320, 32);
-            // Delivery 10
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 288, 320, 1);
-            // Delivery 8
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 320, 384, 64);
-            // Delivery 9
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 320, 352, 32);
-            // Delivery 10
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 320, 352, 1);
-            // Delivery 9
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 352, 384, 32);
-            // Delivery 10
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 352, 384, 1);
-            // Delivery 7
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 384, 512, 128);
-            // Delivery 8
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 384, 448, 64);
-            // Delivery 9
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 384, 416, 32);
-            // Delivery 10
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 384, 416, 1);
-            // Delivery 9
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 416, 448, 32);
-            // Delivery 10
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 416, 448, 1);
-            // Delivery 8
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 448, 512, 64);
-            // Delivery 9
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 448, 480, 32);
-            // Delivery 10
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 448, 480, 1);
-            // Delivery 9
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 480, 512, 32);
-            // Delivery 10
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 480, 512, 1);
-            // Next poll should not have any records as all records have reached delivery limit.
-            ConsumerRecords<byte[], byte[]> records = shareConsumer.poll(Duration.ofMillis(2500L));
-            assertTrue(records.isEmpty(), "Records should be empty as all records have reached delivery limit. But received: " + records.count());
-        }
-    }
-
-    @ClusterTest(
-        serverProperties = {
-            @ClusterConfigProperty(key = "group.share.delivery.count.limit", value = "10"),
-        }
-    )
-    public void testFetchWithThrottledDeliveryValidateDeliveryCount() throws InterruptedException {
-        alterShareAutoOffsetReset("group1", "earliest");
-        try (Producer<byte[], byte[]> producer = createProducer();
-            ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer(
-                "group1",
-                Map.of(ConsumerConfig.SHARE_ACKNOWLEDGEMENT_MODE_CONFIG, EXPLICIT))
-        ) {
-            int producedMessageCount = 500;
-            // Produce a batch of 500 messages
-            for (int i = 0; i < producedMessageCount; i++) {
-                ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(),
-                    null, "key".getBytes(), ("Message " + i).getBytes());
-                producer.send(record);
-            }
-            producer.flush();
-
-            // Map to track delivery count for each offset.
-            Map<Long, Integer> offsetToDeliveryCountMap = new HashMap<>();
-            // Map which defines expected fetch count for each delivery attempt from 1 to 10.
-            Map<Integer, Integer> expectedFetchCountMap = Map.of(1, 500, 2, 500, 3, 500, 4, 500, 5, 500,
-                6, 250, 7, 125, 8, 62, 9, 31, 10, 1);
-            shareConsumer.subscribe(List.of(tp.topic()));
-            // Fetch records in 10 iterations, each time acknowledging with RELEASE. 10 is the
-            // delivery limit hence we should see throttling from Math.ceil(10/2) = 5 fetches.
-            for (int i = 0; i < 10; i++) {
-                int expectedFetchCount = expectedFetchCountMap.get(i + 1);
-                ConsumerRecords<byte[], byte[]> records = waitedPoll(shareConsumer, 2500L, expectedFetchCount);
-                assertEquals(expectedFetchCount, records.count());
-                // Update delivery count for each offset.
-                records.forEach(record -> {
-                    if (!offsetToDeliveryCountMap.containsKey(record.offset())) {
-                        offsetToDeliveryCountMap.put(record.offset(), 1);
-                    } else  {
-                        offsetToDeliveryCountMap.put(record.offset(), offsetToDeliveryCountMap.get(record.offset()) + 1);
-                    }
-                });
-                // Acknowledge with RELEASE.
-                records.forEach(record -> shareConsumer.acknowledge(record, AcknowledgeType.RELEASE));
-                Map<TopicIdPartition, Optional<KafkaException>> result = shareConsumer.commitSync();
-                assertEquals(1, result.size());
-                assertEquals(Optional.empty(), result.get(new TopicIdPartition(tpId, tp.partition(), tp.topic())));
-            }
-
-            // Validate every offset is delivered at most till delivery limit.
-            waitForCondition(() -> {
-                ConsumerRecords<byte[], byte[]> records = shareConsumer.poll(Duration.ofMillis(2500L));
-                if (!records.isEmpty()) {
-                    records.forEach(record -> {
-                        if (!offsetToDeliveryCountMap.containsKey(record.offset())) {
-                            offsetToDeliveryCountMap.put(record.offset(), 1);
-                        } else  {
-                            offsetToDeliveryCountMap.put(record.offset(), offsetToDeliveryCountMap.get(record.offset()) + 1);
-                        }
-                    });
-                    records.forEach(record -> shareConsumer.acknowledge(record, AcknowledgeType.RELEASE));
-                    Map<TopicIdPartition, Optional<KafkaException>> result = shareConsumer.commitSync();
-                    assertEquals(1, result.size());
-                    assertEquals(Optional.empty(),
-                        result.get(new TopicIdPartition(tpId, tp.partition(), tp.topic())));
-                }
-                return offsetToDeliveryCountMap.size() == 500 &&
-                    offsetToDeliveryCountMap.values().stream().allMatch(deliveryCount -> deliveryCount == 10);
-                },
-                120000L, // 120 seconds.
-                50L,
-                () -> "failed to get records till delivery limit"
-            );
-
-            // Next poll should not have any records as all records have reached delivery limit.
-            ConsumerRecords<byte[], byte[]> records = shareConsumer.poll(Duration.ofMillis(2500L));
-            assertTrue(records.isEmpty(), "Records should be empty as all records have reached delivery limit. But received: " + records.count());
-        }
-    }
-
-    @ClusterTest(
-        serverProperties = {
-            @ClusterConfigProperty(key = "group.share.delivery.count.limit", value = "2"),
-        }
-    )
-    public void testFetchWithThrottledDeliveryBatchesWithDecreasedDeliveryLimit() {
-        alterShareAutoOffsetReset("group1", "earliest");
-        try (Producer<byte[], byte[]> producer = createProducer();
-            ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer(
-                "group1",
-                Map.of(
-                    ConsumerConfig.SHARE_ACKNOWLEDGEMENT_MODE_CONFIG, EXPLICIT,
-                    ConsumerConfig.MAX_POLL_RECORDS_CONFIG, 512
-                ))
-        ) {
-            // Produce records in complete power of 2 to fully test the throttling behavior.
-            int producedMessageCount = 512;
-            // Produce a batch of 512 messages
-            for (int i = 0; i < producedMessageCount; i++) {
-                ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(),
-                    null, "key".getBytes(), ("Message " + i).getBytes());
-                producer.send(record);
-            }
-            producer.flush();
-
-            shareConsumer.subscribe(List.of(tp.topic()));
-            // Fetch records in 2 iterations, each time acknowledging with RELEASE. As throttling
-            // currently applies for delivery limit > 2, hence we should get full batch in both fetches.
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 0, 512, 512);
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer, 0, 512, 512);
-            // Next poll should not have any records as all records have reached delivery limit.
-            ConsumerRecords<byte[], byte[]> records = shareConsumer.poll(Duration.ofMillis(2500L));
-            assertTrue(records.isEmpty(), "Records should be empty as all records have reached delivery limit. But received: " + records.count());
-        }
-    }
-
-    @ClusterTest
-    public void testFetchWithThrottledDeliveryBatchesMultipleConsumers() {
-        alterShareAutoOffsetReset("group1", "earliest");
-        try (Producer<byte[], byte[]> producer = createProducer();
-            ShareConsumer<byte[], byte[]> shareConsumer1 = createShareConsumer(
-                "group1",
-                Map.of(
-                    ConsumerConfig.SHARE_ACKNOWLEDGEMENT_MODE_CONFIG, EXPLICIT,
-                    ConsumerConfig.MAX_POLL_RECORDS_CONFIG, 1
-                ));
-            ShareConsumer<byte[], byte[]> shareConsumer2 = createShareConsumer(
-                "group1",
-                Map.of(
-                    ConsumerConfig.SHARE_ACKNOWLEDGEMENT_MODE_CONFIG, EXPLICIT,
-                    ConsumerConfig.MAX_POLL_RECORDS_CONFIG, 2
-                ))
-        ) {
-            // Produce 2 records in separate batches.
-            for (int i = 0; i < 2; i++) {
-                ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(),
-                    null, "key".getBytes(), ("Message " + i).getBytes());
-                producer.send(record);
-                // Flush immediately to create 2 different batches.
-                producer.flush();
-            }
-
-            shareConsumer1.subscribe(List.of(tp.topic()));
-            shareConsumer2.subscribe(List.of(tp.topic()));
-            // Fetch from consumer1 - should get 1 record as max.poll.records=1.
-            ConsumerRecords<byte[], byte[]> records = waitedPoll(shareConsumer1, 2500L, 1);
-            assertEquals(1, records.count());
-            // Verify the first offset of the fetched records.
-            assertEquals(0, records.iterator().next().offset());
-            // Fetch from consumer2 - should get 1 record as offset 0 is Acquired by consumer1.
-            // Release the record from consumer2 after fetching until the last delivery attempt.
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer2, 1, 2, 1);
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer2, 1, 2, 1);
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer2, 1, 2, 1);
-            validateExpectedRecordsInEachPollAndRelease(shareConsumer2, 1, 2, 1);
-
-            // Now release the record from consumer1. Fetch again from consumer2 to verify it gets the released record.
-            // And should only get 1 record at offset 0 as offset 1 record is in final delivery attempt.
-            records.forEach(record -> shareConsumer1.acknowledge(record, AcknowledgeType.RELEASE));
-            Map<TopicIdPartition, Optional<KafkaException>> result = shareConsumer1.commitSync();
-            assertEquals(1, result.size());
-            assertEquals(Optional.empty(), result.get(new TopicIdPartition(tpId, tp.partition(), tp.topic())));
-            // Fetch from consumer2 - should get the released record at offset 0. Accept the record after fetching.
-            validateExpectedRecordsInEachPollAndAcknowledge(shareConsumer2, 0, 1, 1, AcknowledgeType.ACCEPT);
-            // Now fetch the last record at offset 1 from consumer2 in its final delivery attempt.
-            validateExpectedRecordsInEachPollAndAcknowledge(shareConsumer2, 1, 2, 1, AcknowledgeType.ACCEPT);
-
-            // Next poll from consumer1 should not have any records as all records have reached delivery limit.
-            records = shareConsumer1.poll(Duration.ofMillis(2500L));
-            assertTrue(records.isEmpty(), "Records should be empty as all records have reached delivery limit. But received: " + records.count());
-        }
-    }
-
-    @ClusterTest
-    public void testDynamicDeliveryCountLimitDecreaseArchivesRecords() {
-        alterShareAutoOffsetReset("group1", "earliest");
-        try (Producer<byte[], byte[]> producer = createProducer();
-            ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer(
-                "group1",
-                Map.of(ConsumerConfig.SHARE_ACKNOWLEDGEMENT_MODE_CONFIG, EXPLICIT))
-        ) {
-            // Produce 1 record.
-            ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(),
-                null, "key".getBytes(), "value".getBytes());
-            producer.send(record);
-            producer.flush();
-
-            shareConsumer.subscribe(Set.of(tp.topic()));
-
-            // Consume and release twice (deliveryCount becomes 2).
-            // Delivery 1: acquire (deliveryCount=1) → release (1 < default limit 5 → AVAILABLE).
-            ConsumerRecords<byte[], byte[]> records = waitedPoll(shareConsumer, 2500L, 1);
-            assertEquals(1, records.count());
-            records.forEach(r -> shareConsumer.acknowledge(r, AcknowledgeType.RELEASE));
-            shareConsumer.commitSync();
-
-            // Delivery 2: acquire (deliveryCount=2) → release (2 < 5 → AVAILABLE).
-            records = waitedPoll(shareConsumer, 2500L, 1);
-            assertEquals(1, records.count());
-            assertEquals((short) 2, records.records(tp).get(0).deliveryCount().get());
-            records.forEach(r -> shareConsumer.acknowledge(r, AcknowledgeType.RELEASE));
-            shareConsumer.commitSync();
-
-            // Dynamically decrease delivery count limit to 3 via group config.
-            alterShareDeliveryCountLimit("group1", "3");
-
-            // Delivery 3: acquire (deliveryCount=3) → release (3 >= 3 → ARCHIVED).
-            records = waitedPoll(shareConsumer, 2500L, 1);
-            assertEquals(1, records.count());
-            assertEquals((short) 3, records.records(tp).get(0).deliveryCount().get());
-            records.forEach(r -> shareConsumer.acknowledge(r, AcknowledgeType.RELEASE));
-            shareConsumer.commitSync();
-
-            // Next poll should have no records because the record was archived.
-            records = shareConsumer.poll(Duration.ofMillis(2500L));
-            assertTrue(records.isEmpty(),
-                "Records should be empty as the record was archived. But received: " + records.count());
-        }
-    }
-
-    @ClusterTest(
-        serverProperties = {
-            @ClusterConfigProperty(key = "group.share.delivery.count.limit", value = "3"),
-        }
-    )
-    public void testDynamicDeliveryCountLimitIncreaseAllowsMoreDeliveries() {
-        alterShareAutoOffsetReset("group1", "earliest");
-        try (Producer<byte[], byte[]> producer = createProducer();
-            ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer(
-                "group1",
-                Map.of(ConsumerConfig.SHARE_ACKNOWLEDGEMENT_MODE_CONFIG, EXPLICIT))
-        ) {
-            // Produce 1 record.
-            ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(),
-                null, "key".getBytes(), "value".getBytes());
-            producer.send(record);
-            producer.flush();
-
-            shareConsumer.subscribe(Set.of(tp.topic()));
-
-            // Consume and release twice (deliveryCount becomes 2).
-            // Delivery 1: acquire (deliveryCount=1) → release (1 < broker limit 3 → AVAILABLE).
-            ConsumerRecords<byte[], byte[]> records = waitedPoll(shareConsumer, 2500L, 1);
-            assertEquals(1, records.count());
-            records.forEach(r -> shareConsumer.acknowledge(r, AcknowledgeType.RELEASE));
-            shareConsumer.commitSync();
-
-            // Delivery 2: acquire (deliveryCount=2) → release (2 < 3 → AVAILABLE).
-            records = waitedPoll(shareConsumer, 2500L, 1);
-            assertEquals(1, records.count());
-            assertEquals((short) 2, records.records(tp).get(0).deliveryCount().get());
-            records.forEach(r -> shareConsumer.acknowledge(r, AcknowledgeType.RELEASE));
-            shareConsumer.commitSync();
-
-            // Dynamically increase delivery count limit to 5 via group config.
-            alterShareDeliveryCountLimit("group1", "5");
-
-            // Delivery 3: acquire (deliveryCount=3) → release (3 < 5 → AVAILABLE).
-            // Without the config increase, 3 >= 3 would have caused archival.
-            records = waitedPoll(shareConsumer, 2500L, 1);
-            assertEquals(1, records.count());
-            assertEquals((short) 3, records.records(tp).get(0).deliveryCount().get());
-            records.forEach(r -> shareConsumer.acknowledge(r, AcknowledgeType.RELEASE));
-            shareConsumer.commitSync();
-
-            // Delivery 4: acquire (deliveryCount=4) → release (4 < 5 → AVAILABLE).
-            records = waitedPoll(shareConsumer, 2500L, 1);
-            assertEquals(1, records.count());
-            assertEquals((short) 4, records.records(tp).get(0).deliveryCount().get());
-            records.forEach(r -> shareConsumer.acknowledge(r, AcknowledgeType.RELEASE));
-            shareConsumer.commitSync();
-
-            // Delivery 5: acquire (deliveryCount=5) → release (5 >= 5 → ARCHIVED).
-            records = waitedPoll(shareConsumer, 2500L, 1);
-            assertEquals(1, records.count());
-            assertEquals((short) 5, records.records(tp).get(0).deliveryCount().get());
-            records.forEach(r -> shareConsumer.acknowledge(r, AcknowledgeType.RELEASE));
-            shareConsumer.commitSync();
-
-            // Next poll should have no records because the record was archived.
-            records = shareConsumer.poll(Duration.ofMillis(2500L));
-            assertTrue(records.isEmpty(),
-                "Records should be empty as the record was archived. But received: " + records.count());
-        }
-    }
-
-    @ClusterTest
-    public void testDynamicPartitionMaxRecordLocks() {
-        // Verify that the group-level share.partition.max.record.locks config can be
-        // dynamically set and read back via describe configs.
-        alterSharePartitionMaxRecordLocks("group1", "500");
-
-        // Verify the config can be updated dynamically.
-        alterSharePartitionMaxRecordLocks("group1", "1000");
-    }
-
-    /**
-     * Util class to encapsulate state for a consumer/producer
-     * being executed by an {@link ExecutorService}.
-     */
-    private static class ClientState {
-        private final AtomicBoolean done = new AtomicBoolean(false);
-        private final AtomicInteger count = new AtomicInteger(0);
-
-        AtomicBoolean done() {
-            return done;
-        }
-
-        AtomicInteger count() {
-            return count;
-        }
-    }
-
-    private int produceMessages(int messageCount) {
-        try (Producer<byte[], byte[]> producer = createProducer()) {
-            ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(), null, "key".getBytes(), "value".getBytes());
-            IntStream.range(0, messageCount).forEach(__ -> producer.send(record));
-            producer.flush();
-        }
-        return messageCount;
-    }
-
-    private void produceMessagesWithTimestamp(int messageCount, long startingTimestamp) {
-        try (Producer<byte[], byte[]> producer = createProducer()) {
-            for (int i = 0; i < messageCount; i++) {
-                ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(), startingTimestamp + i,
-                    ("key " + i).getBytes(), ("value " + i).getBytes());
-                producer.send(record);
-            }
-            producer.flush();
-        }
-    }
-
-    private void produceCommittedTransaction(Producer<byte[], byte[]> transactionalProducer, String message) {
-        try {
-            transactionalProducer.beginTransaction();
-            ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(), null, message.getBytes(), message.getBytes());
-            Future<RecordMetadata> future = transactionalProducer.send(record);
-            transactionalProducer.flush();
-            future.get(); // Ensure producer send is complete before committing
-            transactionalProducer.commitTransaction();
-        } catch (Exception e) {
-            transactionalProducer.abortTransaction();
-        }
-    }
-
-    private void produceAbortedTransaction(Producer<byte[], byte[]> transactionalProducer, String message) {
-        try {
-            transactionalProducer.beginTransaction();
-            ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(), null, message.getBytes(), message.getBytes());
-            Future<RecordMetadata> future = transactionalProducer.send(record);
-            transactionalProducer.flush();
-            future.get(); // Ensure producer send is complete before aborting
-            transactionalProducer.abortTransaction();
-        } catch (Exception e) {
-            transactionalProducer.abortTransaction();
-        }
-    }
-
-    private void produceCommittedAndAbortedTransactionsInInterval(Producer<byte[], byte[]> transactionalProducer, int messageCount, int intervalAbortedTransactions) {
-        transactionalProducer.initTransactions();
-        int transactionCount = 0;
-        try {
-            for (int i = 0; i < messageCount; i++) {
-                transactionalProducer.beginTransaction();
-                String recordMessage = "Message " + (i + 1);
-                ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(tp.topic(), tp.partition(), null, recordMessage.getBytes(), recordMessage.getBytes());
-                Future<RecordMetadata> future = transactionalProducer.send(record);
-                transactionalProducer.flush();
-                // Increment transaction count
-                transactionCount++;
-                if (transactionCount % intervalAbortedTransactions == 0) {
-                    // Aborts every intervalAbortedTransactions transaction
-                    transactionalProducer.abortTransaction();
-                } else {
-                    // Commits other transactions
-                    future.get(); // Ensure producer send is complete before committing
-                    transactionalProducer.commitTransaction();
-                }
-            }
-        } catch (Exception e) {
-            transactionalProducer.abortTransaction();
-        } finally {
-            transactionalProducer.close();
-        }
-    }
-
-    private int consumeMessages(AtomicInteger totalMessagesConsumed,
-                                int totalMessages,
-                                String groupId,
-                                int consumerNumber,
-                                int maxPolls,
-                                boolean commit) {
-        return assertDoesNotThrow(() -> {
-            try (ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer(groupId)) {
-                shareConsumer.subscribe(Set.of(tp.topic()));
-                return consumeMessages(shareConsumer, totalMessagesConsumed, totalMessages, consumerNumber, maxPolls, commit);
-            }
-        }, "Consumer " + consumerNumber + " failed with exception");
-    }
-
-    private int consumeMessages(AtomicInteger totalMessagesConsumed,
-                                int totalMessages,
-                                String groupId,
-                                int consumerNumber,
-                                int maxPolls,
-                                boolean commit,
-                                int maxFetchBytes) {
-        return assertDoesNotThrow(() -> {
-            try (ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer(
-                groupId,
-                Map.of(ConsumerConfig.MAX_PARTITION_FETCH_BYTES_CONFIG, maxFetchBytes))) {
-                shareConsumer.subscribe(Set.of(tp.topic()));
-                return consumeMessages(shareConsumer, totalMessagesConsumed, totalMessages, consumerNumber, maxPolls, commit);
-            }
-        }, "Consumer " + consumerNumber + " failed with exception");
-    }
-
-    private int consumeMessages(ShareConsumer<byte[], byte[]> consumer,
-                                AtomicInteger totalMessagesConsumed,
-                                int totalMessages,
-                                int consumerNumber,
-                                int maxPolls,
-                                boolean commit) {
-        return assertDoesNotThrow(() -> {
-            int messagesConsumed = 0;
-            int retries = 0;
-            if (totalMessages > 0) {
-                while (totalMessagesConsumed.get() < totalMessages && retries < maxPolls) {
-                    ConsumerRecords<byte[], byte[]> records = consumer.poll(Duration.ofMillis(2000));
-                    messagesConsumed += records.count();
-                    totalMessagesConsumed.addAndGet(records.count());
-                    retries++;
-                }
-            } else {
-                while (retries < maxPolls) {
-                    ConsumerRecords<byte[], byte[]> records = consumer.poll(Duration.ofMillis(2000));
-                    messagesConsumed += records.count();
-                    totalMessagesConsumed.addAndGet(records.count());
-                    retries++;
-                }
-            }
-
-            if (commit) {
-                // Complete acknowledgement of the records
-                consumer.commitSync(Duration.ofMillis(10000));
-            }
-            return messagesConsumed;
-        }, "Consumer " + consumerNumber + " failed with exception");
-    }
-
-    private <K, V> List<ConsumerRecord<K, V>> consumeRecords(ShareConsumer<K, V> consumer,
-                                                             int numRecords) {
-        ArrayList<ConsumerRecord<K, V>> accumulatedRecords = new ArrayList<>();
-        long startTimeMs = System.currentTimeMillis();
-        while (accumulatedRecords.size() < numRecords) {
-            ConsumerRecords<K, V> records = consumer.poll(Duration.ofMillis(100));
-            records.forEach(accumulatedRecords::add);
-            long currentTimeMs = System.currentTimeMillis();
-            assertFalse(currentTimeMs - startTimeMs > 60000, "Timed out before consuming expected records.");
-        }
-        return accumulatedRecords;
-    }
-
-    private Uuid createTopic(String topicName) {
-        return createTopic(topicName, 1, 1);
-    }
-
-    private Uuid createTopic(String topicName, int numPartitions, int replicationFactor) {
-        AtomicReference<Uuid> topicId = new AtomicReference<>(null);
-        assertDoesNotThrow(() -> {
-            try (Admin admin = createAdminClient()) {
-                CreateTopicsResult result = admin.createTopics(Set.of(new NewTopic(topicName, numPartitions, (short) replicationFactor)));
-                result.all().get();
-                topicId.set(result.topicId(topicName).get());
-            }
-        }, "Failed to create topic");
-
-        return topicId.get();
-    }
-
-    private void deleteTopic(String topicName) {
-        assertDoesNotThrow(() -> {
-            try (Admin admin = createAdminClient()) {
-                admin.deleteTopics(Set.of(topicName)).all().get();
-            }
-        }, "Failed to delete topic");
-    }
-
-    private Admin createAdminClient() {
-        return cluster.admin();
-    }
-
-    private <K, V> Producer<K, V> createProducer() {
-        return createProducer(Map.of());
-    }
-
-    private <K, V> Producer<K, V> createProducer(Map<String, Object> config) {
-        return cluster.producer(config);
-    }
-
-    private <K, V> Producer<K, V> createProducer(String transactionalId) {
-        return createProducer(
-            Map.of(
-                ProducerConfig.TRANSACTIONAL_ID_CONFIG, transactionalId
-            )
-        );
-    }
-
-    private <K, V> ShareConsumer<K, V> createShareConsumer(String groupId) {
-        return createShareConsumer(groupId, Map.of());
-    }
-
-    private <K, V> ShareConsumer<K, V> createShareConsumer(
-        String groupId,
-        Map<?, ?> additionalProperties
-    ) {
-        return createShareConsumer(groupId, additionalProperties, null, null);
-    }
-
-    private <K, V> ShareConsumer<K, V> createShareConsumer(
-        String groupId,
-        Map<?, ?> additionalProperties,
-        Deserializer<K> keyDeserializer,
-        Deserializer<V> valueDeserializer
-    ) {
-        Properties props = new Properties();
-        props.putAll(additionalProperties);
-        props.put(ConsumerConfig.GROUP_ID_CONFIG, groupId);
-        Map<String, Object> conf = new HashMap<>();
-        props.forEach((k, v) -> conf.put((String) k, v));
-        return cluster.shareConsumer(conf, keyDeserializer, valueDeserializer);
-    }
-
-    private void warmup() throws InterruptedException {
-        createTopic(warmupTp.topic());
-        waitForMetadataCache();
-        ProducerRecord<byte[], byte[]> record = new ProducerRecord<>(warmupTp.topic(), warmupTp.partition(), null, "key".getBytes(), "value".getBytes());
-        Set<String> subscription = Set.of(warmupTp.topic());
-        alterShareAutoOffsetReset("warmupgroup1", "earliest");
-        try (Producer<byte[], byte[]> producer = createProducer();
-             ShareConsumer<byte[], byte[]> shareConsumer = createShareConsumer("warmupgroup1")) {
-
-            producer.send(record);
-            producer.flush();
-
-            shareConsumer.subscribe(subscription);
-            TestUtils.waitForCondition(
-                () -> shareConsumer.poll(Duration.ofMillis(5000)).count() == 1, 30000, 200L, () -> "warmup record not received");
-        }
-    }
-
-    private void waitForMetadataCache() throws InterruptedException {
-        TestUtils.waitForCondition(() ->
-                !cluster.brokers().get(0).metadataCache().getAliveBrokerNodes(new ListenerName("EXTERNAL")).isEmpty(),
-            DEFAULT_MAX_WAIT_MS, 100L, () -> "cache not up yet");
-    }
-
-    private void verifyShareGroupStateTopicRecordsProduced() {
-        try {
-            try (Consumer<byte[], byte[]> consumer = cluster.consumer()) {
-                consumer.assign(sgsTopicPartitions);
-                consumer.seekToBeginning(sgsTopicPartitions);
-                Set<ConsumerRecord<byte[], byte[]>> records = new HashSet<>();
-                TestUtils.waitForCondition(() -> {
-                        ConsumerRecords<byte[], byte[]> msgs = consumer.poll(Duration.ofMillis(5000L));
-                        if (msgs.count() > 0) {
-                            msgs.records(Topic.SHARE_GROUP_STATE_TOPIC_NAME).forEach(records::add);
-                        }
-                        return records.size() > 2; // +2 because of extra warmup records
-                    },
-                    30000L,
-                    200L,
-                    () -> "no records produced"
-                );
-            }
-        } catch (InterruptedException e) {
-            fail(e);
-        }
-    }
-
-    private void alterShareGroupConfig(String groupId, String configKey, String newValue) {
-        ConfigResource configResource = new ConfigResource(ConfigResource.Type.GROUP, groupId);
-        Map<ConfigResource, Collection<AlterConfigOp>> alterEntries = new HashMap<>();
-        alterEntries.put(configResource, List.of(new AlterConfigOp(new ConfigEntry(
-            configKey, newValue), AlterConfigOp.OpType.SET)));
-        AlterConfigsOptions alterOptions = new AlterConfigsOptions();
-        try (Admin adminClient = createAdminClient()) {
-            assertDoesNotThrow(() -> adminClient.incrementalAlterConfigs(alterEntries, alterOptions)
-                .all()
-                .get(60, TimeUnit.SECONDS), "Failed to alter configs");
-        }
-
-        // This config is changed dynamically in tests, and we need it to have propagated before the test proceeds.
-        // Describing the config with a new admin client is not totally foolproof, but it's better than just
-        // altering the config and continuing.
-        try (Admin adminClient = createAdminClient()) {
-            assertDoesNotThrow(() ->
-                TestUtils.waitForCondition(() -> {
-                    Config config = adminClient.describeConfigs(List.of(configResource)).all().get().get(configResource);
-                    ConfigEntry entry = config.get(configKey);
-                    return entry != null && entry.value().equals(newValue);
-                }, 10000L, 100L, () -> "New config value did not propagate"), "Failed to describe configs");
-        }
-    }
-
-    private void alterShareAutoOffsetReset(String groupId, String newValue) {
-        alterShareGroupConfig(groupId, GroupConfig.SHARE_AUTO_OFFSET_RESET_CONFIG, newValue);
-    }
-
-    private void alterShareDeliveryCountLimit(String groupId, String newValue) {
-        alterShareGroupConfig(groupId, GroupConfig.SHARE_DELIVERY_COUNT_LIMIT_CONFIG, newValue);
-    }
-
-    private void alterShareIsolationLevel(String groupId, String newValue) {
-        alterShareGroupConfig(groupId, GroupConfig.SHARE_ISOLATION_LEVEL_CONFIG, newValue);
-    }
-
-    private void alterShareRenewAcknowledgeEnable(String groupId, boolean newValue) {
-        alterShareGroupConfig(groupId, GroupConfig.SHARE_RENEW_ACKNOWLEDGE_ENABLE_CONFIG, Boolean.toString(newValue));
-    }
-
-    private List<Integer> topicPartitionLeader(Admin adminClient, String topicName, int partition) throws InterruptedException, ExecutionException {
-        return adminClient.describeTopics(List.of(topicName)).allTopicNames().get().get(topicName)
-            .partitions().stream()
-            .filter(info -> info.partition() == partition)
-            .map(info -> info.leader().id())
-            .filter(info -> info != -1)
-            .toList();
-    }
-
-    private SharePartitionOffsetInfo sharePartitionOffsetInfo(Admin adminClient, String groupId, TopicPartition tp) throws InterruptedException, ExecutionException {
-        SharePartitionOffsetInfo partitionResult;
-        ListShareGroupOffsetsResult result = adminClient.listShareGroupOffsets(
-            Map.of(groupId, new ListShareGroupOffsetsSpec().topicPartitions(List.of(tp))),
-            new ListShareGroupOffsetsOptions().timeoutMs(30000)
-        );
-        partitionResult = result.partitionsToOffsetInfo(groupId).get().get(tp);
-        return partitionResult;
-    }
-
-    private void verifySharePartitionStartOffset(Admin adminClient, String groupId, TopicPartition tp, long expectedStartOffset) throws InterruptedException {
-        TestUtils.waitForCondition(() -> {
-            SharePartitionOffsetInfo sharePartitionOffsetInfo = sharePartitionOffsetInfo(adminClient, groupId, tp);
-            return sharePartitionOffsetInfo != null &&
-                sharePartitionOffsetInfo.startOffset() == expectedStartOffset;
-        }, DEFAULT_MAX_WAIT_MS, DEFAULT_POLL_INTERVAL_MS, () -> "Failed to retrieve share partition lag");
-    }
-
-    private void verifySharePartitionLag(Admin adminClient, String groupId, TopicPartition tp, long expectedLag) throws InterruptedException {
-        TestUtils.waitForCondition(() -> {
-            SharePartitionOffsetInfo sharePartitionOffsetInfo = sharePartitionOffsetInfo(adminClient, groupId, tp);
-            return sharePartitionOffsetInfo != null &&
-                sharePartitionOffsetInfo.lag().isPresent() &&
-                sharePartitionOffsetInfo.lag().get() == expectedLag;
-        }, DEFAULT_MAX_WAIT_MS, DEFAULT_POLL_INTERVAL_MS, () -> "Failed to retrieve share partition lag");
-    }
-
-    private void verifySharePartitionOffsetsDeleted(Admin adminClient, String groupId, TopicPartition tp) throws InterruptedException {
-        TestUtils.waitForCondition(
-            () -> sharePartitionOffsetInfo(adminClient, groupId, tp) == null, 
-            DEFAULT_MAX_WAIT_MS, 
-            DEFAULT_POLL_INTERVAL_MS, 
-            () -> "Failed to retrieve share partition lag");
-    }
-
-    private void alterShareGroupOffsets(Admin adminClient, String groupId, TopicPartition topicPartition, Long newOffset) throws InterruptedException, ExecutionException {
-        adminClient.alterShareGroupOffsets(
-            groupId,
-            Map.of(topicPartition, newOffset),
-            new AlterShareGroupOffsetsOptions().timeoutMs(30000)).partitionResult(topicPartition).get();
-    }
-
-    private void deleteShareGroupOffsets(Admin adminClient, String groupId, String topic) throws InterruptedException, ExecutionException {
-        adminClient.deleteShareGroupOffsets(
-            groupId,
-            Set.of(topic),
-            new DeleteShareGroupOffsetsOptions().timeoutMs(30000)).topicResult(topic).get();
-    }
-
-    private void alterSharePartitionMaxRecordLocks(String groupId, String newValue) {
-        alterShareGroupConfig(groupId, GroupConfig.SHARE_PARTITION_MAX_RECORD_LOCKS_CONFIG, newValue);
-    }
-
-    private void alterShareRecordLockDurationMs(String groupId, int newValue) {
-        alterShareGroupConfig(groupId, GroupConfig.SHARE_RECORD_LOCK_DURATION_MS_CONFIG, Integer.toString(newValue));
-    }
-
-    /**
-     * Test utility which encapsulates a {@link ShareConsumer} whose record processing
-     * behavior can be supplied as a function argument.
-     * <p></p>
-     * This can be used to create different consume patterns on the broker and study
-     * the status of broker side share group abstractions.
-     *
-     * @param <K> - key type of the records consumed
-     * @param <V> - value type of the records consumed
-     */
-    private static class ComplexShareConsumer<K, V> implements Runnable {
-        public static final int POLL_TIMEOUT_MS = 15000;
-        public static final int MAX_DELIVERY_COUNT = ShareGroupConfig.SHARE_GROUP_DELIVERY_COUNT_LIMIT_DEFAULT;
-
-        private final String topicName;
-        private final Map<String, Object> configs = new HashMap<>();
-        private final ClientState state = new ClientState();
-        private final Predicate<ConsumerRecords<K, V>> exitCriteria;
-        private final BiConsumer<ShareConsumer<K, V>, ConsumerRecord<K, V>> processFunc;
-
-        ComplexShareConsumer(
-            String bootstrapServers,
-            String topicName,
-            String groupId,
-            Map<String, Object> additionalProperties
-        ) {
-            this(
-                bootstrapServers,
-                topicName,
-                groupId,
-                additionalProperties,
-                records -> records.count() == 0,
-                (consumer, record) -> {
-                    short deliveryCountBeforeAccept = (short) ((record.offset() + record.offset() / (MAX_DELIVERY_COUNT + 2)) % (MAX_DELIVERY_COUNT + 2));
-                    if (deliveryCountBeforeAccept == 0) {
-                        consumer.acknowledge(record, AcknowledgeType.REJECT);
-                    } else if (record.deliveryCount().get() == deliveryCountBeforeAccept) {
-                        consumer.acknowledge(record, AcknowledgeType.ACCEPT);
-                    } else {
-                        consumer.acknowledge(record, AcknowledgeType.RELEASE);
-                    }
-                }
-            );
-        }
-
-        ComplexShareConsumer(
-            String bootstrapServers,
-            String topicName,
-            String groupId,
-            Map<String, Object> additionalProperties,
-            Predicate<ConsumerRecords<K, V>> exitCriteria,
-            BiConsumer<ShareConsumer<K, V>, ConsumerRecord<K, V>> processFunc
-        ) {
-            this.exitCriteria = Objects.requireNonNull(exitCriteria);
-            this.processFunc = Objects.requireNonNull(processFunc);
-            this.topicName = topicName;
-            this.configs.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapServers);
-            this.configs.put(ConsumerConfig.GROUP_ID_CONFIG, groupId);
-            this.configs.putAll(additionalProperties);
-            this.configs.putIfAbsent(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, ByteArrayDeserializer.class);
-            this.configs.putIfAbsent(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, ByteArrayDeserializer.class);
-        }
-
-        @Override
-        public void run() {
-            try (ShareConsumer<K, V> consumer = new KafkaShareConsumer<>(configs)) {
-                consumer.subscribe(Set.of(this.topicName));
-                while (!state.done().get()) {
-                    ConsumerRecords<K, V> records = consumer.poll(Duration.ofMillis(POLL_TIMEOUT_MS));
-                    state.count().addAndGet(records.count());
-                    if (exitCriteria.test(records)) {
-                        state.done().set(true);
-                    }
-                    records.forEach(record -> processFunc.accept(consumer, record));
-                }
-            }
-        }
-
-        boolean isDone() {
-            return state.done().get();
-        }
-
-        int recordsRead() {
-            return state.count().get();
-        }
-    }
-
-    private void shutdownExecutorService(ExecutorService service) {
-        service.shutdown();
-        try {
-            if (!service.awaitTermination(5L, TimeUnit.SECONDS)) {
-                service.shutdownNow();
-            }
-        } catch (Exception e) {
-            service.shutdownNow();
-            Thread.currentThread().interrupt();
-        }
-    }
-
-    private ConsumerRecords<byte[], byte[]> waitedPoll(
-        ShareConsumer<byte[], byte[]> shareConsumer,
-        long pollMs,
-        int recordCount
-    ) {
-        return waitedPoll(shareConsumer, pollMs, recordCount, false, "", List.of());
-    }
-
-    private ConsumerRecords<byte[], byte[]> waitedPoll(
-        ShareConsumer<byte[], byte[]> shareConsumer,
-        long pollMs,
-        int recordCount,
-        boolean checkAssignment,
-        String groupId,
-        List<TopicPartition> tps
-    ) {
-        AtomicReference<ConsumerRecords<byte[], byte[]>> recordsAtomic = new AtomicReference<>();
-        try {
-            waitForCondition(() -> {
-                    ConsumerRecords<byte[], byte[]> recs = shareConsumer.poll(Duration.ofMillis(pollMs));
-                    recordsAtomic.set(recs);
-                    if (checkAssignment) {
-                        waitForAssignment(groupId, tps);
-                    }
-                    return recs.count() == recordCount;
-                },
-                DEFAULT_MAX_WAIT_MS,
-                500L,
-                () -> "failed to get records"
-            );
-            return recordsAtomic.get();
-        } catch (InterruptedException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    private void validateExpectedRecordsInEachPollAndRelease(
-        ShareConsumer<byte[], byte[]> shareConsumer,
-        int startOffset,
-        int lastOffset,
-        int expectedRecordsInEachPoll
-    ) {
-        validateExpectedRecordsInEachPollAndAcknowledge(shareConsumer, startOffset, lastOffset, expectedRecordsInEachPoll, AcknowledgeType.RELEASE);
-    }
-
-    private void validateExpectedRecordsInEachPollAndAcknowledge(
-        ShareConsumer<byte[], byte[]> shareConsumer,
-        int startOffset,
-        int lastOffset,
-        int expectedRecordsInEachPoll,
-        AcknowledgeType acknowledgeType
-    ) {
-        for (int i = startOffset; i < lastOffset; i = i + expectedRecordsInEachPoll) {
-            ConsumerRecords<byte[], byte[]> records = waitedPoll(shareConsumer, 2500L, expectedRecordsInEachPoll);
-            assertEquals(expectedRecordsInEachPoll, records.count());
-            // Verify the first offset of the fetched records.
-            assertEquals(i, records.iterator().next().offset());
-
-            records.forEach(record -> shareConsumer.acknowledge(record, acknowledgeType));
-            Map<TopicIdPartition, Optional<KafkaException>> result = shareConsumer.commitSync();
-            assertEquals(1, result.size());
-            assertEquals(Optional.empty(), result.get(new TopicIdPartition(tpId, tp.partition(), tp.topic())));
-        }
-    }
-
-    private void waitForAssignment(String groupId, List<TopicPartition> tps) {
-        try {
-            waitForCondition(() -> {
-                    try (Admin admin = createAdminClient()) {
-                        Collection<ShareMemberDescription> members = admin.describeShareGroups(List.of(groupId),
-                            new DescribeShareGroupsOptions().includeAuthorizedOperations(true)
-                        ).describedGroups().get(groupId).get().members();
-                        Set<TopicPartition> assigned = new HashSet<>();
-                        members.forEach(desc -> {
-                            if (desc.assignment() != null) {
-                                assigned.addAll(desc.assignment().topicPartitions());
-                            }
-                        });
-                        return assigned.containsAll(tps);
-                    } catch (Exception e) {
-                        throw new RuntimeException(e);
-                    }
-                },
-                DEFAULT_MAX_WAIT_MS,
-                1000L,
-                () -> "tps not assigned to members"
-            );
-        } catch (Exception e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    public static class SerializerImpl implements Serializer<byte[]> {
-
-        @Override
-        public byte[] serialize(String topic, Headers headers, byte[] data) {
-            headers.add(KEY, VALUE.getBytes());
-            return data;
-        }
-
-        @Override
-        public byte[] serialize(String topic, byte[] data) {
-            fail("method should not be invoked");
-            return null;
-        }
-    }
-
-    public static class DeserializerImpl implements Deserializer<byte[]> {
-
-        @Override
-        public byte[] deserialize(String topic, Headers headers, byte[] data) {
-            Header header = headers.lastHeader(KEY);
-            assertEquals("application/octet-stream", header == null ? null : new String(header.value()));
-            return data;
-        }
-
-        @Override
-        public byte[] deserialize(String topic, byte[] data) {
-            fail("method should not be invoked");
-            return null;
         }
     }
 }
